refactor(daemon): split main.rs into focused modules
Pure code motion along the seams the daemon grew organically: context,
startup, factory, injection, pipeline, dictation, command_mode,
selection, speak and notify. main.rs keeps argument parsing, startup
wiring and the IPC dispatch.

Part of #87.
