+++ src/daemon/main.rs
@@ -1,715 +1,38 @@
-use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::{Arc, Mutex as StdMutex, OnceLock};
-use std::time::Duration;
+use std::sync::Arc;
 
 use anyhow::{Context, Result};
 use clap::Parser;
 use tokio::io::AsyncWriteExt;
 use tokio::net::UnixListener;
 use tokio::sync::Mutex;
-use tracing::{debug, error, info, warn};
+use tracing::{error, info};
 
-use audio_silence_gate::{audio_gate_reason, AutoStopDetector, SILENCE_RMS_THRESHOLD};
-use filler_remove::FillerFilter;
-use prompt_echo::is_prompt_echo;
-use whisrs::audio::capture::{AudioCaptureHandle, SAMPLE_RATE};
-use whisrs::audio::feedback;
-use whisrs::history::{self, HistoryEntry};
-use whisrs::llm;
-use whisrs::state::{Action, StateMachine};
-use whisrs::transcription::asr_sidecar::AsrSidecarBackend;
-use whisrs::transcription::deepgram::{DeepgramRestBackend, DeepgramStreamingBackend};
-use whisrs::transcription::groq::GroqBackend;
-use whisrs::transcription::local_parakeet::ParakeetBackend;
-use whisrs::transcription::local_vosk::VoskBackend;
-use whisrs::transcription::local_whisper::LocalWhisperBackend;
-use whisrs::transcription::openai_compatible_realtime::OpenAiCompatibleRealtimeBackend;
-use whisrs::transcription::openai_realtime::OpenAIRealtimeBackend;
-use whisrs::transcription::openai_rest::OpenAIRestBackend;
-use whisrs::transcription::{TranscriptionBackend, TranscriptionConfig};
+use whisrs::history;
 use whisrs::window::{self, WindowTracker};
-use whisrs::{
-    encode_message, read_message, socket_path, validate_language_override, Command, Config,
-    InjectorBackend, LocalWhisperConfig, Response, State,
+use whisrs::{encode_message, read_message, socket_path, Command, Response, State};
+
+mod command_mode;
+mod context;
+mod dictation;
+mod factory;
+mod injection;
+mod notify;
+mod pipeline;
+mod selection;
+mod speak;
+mod startup;
+
+use crate::command_mode::{handle_command_mode, handle_llm_command, handle_set_llm_instruction};
+use crate::context::{DaemonContext, DaemonState};
+use crate::dictation::{handle_cancel, handle_toggle};
+use crate::factory::create_backend;
+use crate::injection::warm_keyboard;
+use crate::notify::send_notification;
+use crate::speak::handle_speak;
+use crate::startup::{
+    check_audio_devices, check_uinput_access, cleanup_stale_socket, import_compositor_env,
+    load_config, validate_config,
 };
-
-static KEYBOARD: OnceLock<StdMutex<Option<Box<dyn xkb_type::KeyInjector>>>> = OnceLock::new();
-const TYPING_DRAIN_TIMEOUT: Duration = Duration::from_secs(5);
-
-/// Context saved when command mode starts recording.
-struct CommandModeContext {
-    selected_text: String,
-    llm_config: llm::LlmConfig,
-}
-
-/// Context saved when a named LLM command (`[[llm_commands]]`) starts recording.
-///
-/// Simpler than [`CommandModeContext`]: there's no selection or clipboard
-/// involved — this is a toggle-recording flavor of plain dictation where the
-/// transcribed text is run through a fixed instruction before typing.
-struct LlmCommandContext {
-    name: String,
-    instruction: String,
-    llm_config: llm::LlmConfig,
-}
-
-/// Shared daemon state protected by a mutex.
-struct DaemonState {
-    state_machine: StateMachine,
-    audio_capture: Option<AudioCaptureHandle>,
-    /// The window that was focused when recording started.
-    recording_window_id: Option<String>,
-    /// Handle to the background streaming pipeline (if active).
-    streaming_task: Option<tokio::task::JoinHandle<Result<String>>>,
-    /// Cancel flag for the active streaming pipeline.
-    ///
-    /// Aborting `streaming_task` only drops the outer pipeline future — the
-    /// spawned backend and typing tasks are detached JoinHandles that keep
-    /// running, and the backend treats the dropped audio channel as a normal
-    /// end-of-stream (it flushes and the typing task types the trailing
-    /// phrase). Setting this flag makes the typing loop discard everything
-    /// instead, so `cancel` never types text. Created per recording.
-    streaming_cancel: Option<Arc<AtomicBool>>,
-    /// When recording started (for duration tracking).
-    recording_started_at: Option<std::time::Instant>,
-    /// The language for the active dictation session, resolved at the
-    /// Idle→Recording transition (per-toggle override or config default).
-    /// Consumed when the session ends so an override never leaks into a
-    /// later recording.
-    session_language: Option<String>,
-    /// Active command mode context (set when command mode is recording).
-    command_mode: Option<CommandModeContext>,
-    /// Active named LLM command context (set when an `[[llm_commands]]`
-    /// hotkey is recording).
-    llm_command: Option<LlmCommandContext>,
-    /// Runtime instruction overrides for `[[llm_commands]]`, keyed by command
-    /// name. Written by the `set_hotkey` path (`SetLlmInstruction`) so a
-    /// reprogrammed instruction takes effect immediately — the loaded
-    /// `Config` is an immutable `Arc`, so this map is the live source of truth
-    /// (and the change is also persisted to disk for the next start).
-    llm_instruction_overrides: std::collections::HashMap<String, String>,
-    /// Stop flag for in-progress TTS playback (read-selection-aloud).
-    ///
-    /// Set when a `Speak` synthesis succeeds and playback begins; cleared when
-    /// playback finishes. `Cancel` and a repeat `Speak` both flip it to `true`
-    /// to interrupt playback. Read-aloud runs independently of the recording
-    /// state machine, so there is no dedicated `State` variant.
-    tts_stop: Option<Arc<std::sync::atomic::AtomicBool>>,
-}
-
-impl DaemonState {
-    fn new() -> Self {
-        Self {
-            state_machine: StateMachine::new(),
-            audio_capture: None,
-            recording_window_id: None,
-            streaming_task: None,
-            streaming_cancel: None,
-            recording_started_at: None,
-            session_language: None,
-            command_mode: None,
-            llm_command: None,
-            llm_instruction_overrides: std::collections::HashMap::new(),
-            tts_stop: None,
-        }
-    }
-
-    /// Consume the active session's language, falling back to the config
-    /// default. Taking (not reading) the value is what ends the language
-    /// session — a per-toggle override applies to exactly one recording.
-    fn take_session_language(&mut self, default_lang: &str) -> String {
-        self.session_language
-            .take()
-            .unwrap_or_else(|| default_lang.to_string())
-    }
-}
-
-/// Resources shared across all connections (not behind the per-request mutex).
-struct DaemonContext {
-    config: Config,
-    window_tracker: Arc<dyn WindowTracker>,
-    transcription_backend: Arc<dyn TranscriptionBackend>,
-    notify: bool,
-    /// Broadcast channel for state changes (consumed by system tray).
-    state_tx: tokio::sync::watch::Sender<State>,
-    /// Normalized microphone level for speech-reactive overlays.
-    overlay_level_tx: Option<tokio::sync::watch::Sender<f32>>,
-    /// `true` when the visual overlay is enabled. State-transition toasts
-    /// are suppressed in that case to avoid duplicate signaling — the
-    /// overlay already shows recording/transcribing state visually.
-    /// Error notifications still fire.
-    overlay_enabled: bool,
-}
-
-impl DaemonContext {
-    /// Should we surface a state/progress notification (not an error)?
-    /// Suppressed when the overlay is on so we don't double-signal the
-    /// same event.
-    fn notify_state(&self) -> bool {
-        self.notify && !self.overlay_enabled
-    }
-
-    /// Should we surface an error notification? Always yes when the user
-    /// has notifications enabled — errors are critical and the overlay
-    /// can't carry their detail.
-    fn notify_error(&self) -> bool {
-        self.notify
-    }
-}
-
-/// Try to connect to an existing socket.
-async fn socket_is_alive(path: &std::path::Path) -> bool {
-    tokio::net::UnixStream::connect(path).await.is_ok()
-}
-
-/// Remove a stale socket file if no daemon is listening on it.
-async fn cleanup_stale_socket(path: &std::path::Path) -> Result<()> {
-    if path.exists() {
-        if socket_is_alive(path).await {
-            anyhow::bail!("another whisrsd instance is already running");
-        }
-        warn!("removing stale socket at {}", path.display());
-        std::fs::remove_file(path).context("failed to remove stale socket")?;
-    }
-    Ok(())
-}
-
-/// Load configuration from config.toml, falling back to defaults.
-/// Returns (Config, Option<warning_message>) — the warning is set when config
-/// parsing fails and defaults are used, so the caller can notify the user.
-fn load_config() -> (Config, Option<String>) {
-    let config_path = whisrs::config_path();
-    if config_path.exists() {
-        match std::fs::read_to_string(&config_path) {
-            Ok(contents) => match toml::from_str::<Config>(&contents) {
-                Ok(config) => {
-                    info!("loaded config from {}", config_path.display());
-                    return (config, None);
-                }
-                Err(e) => {
-                    let msg = format!(
-                        "Failed to parse config at {}: {e} — using defaults",
-                        config_path.display()
-                    );
-                    error!("{msg}");
-                    return (
-                        Config {
-                            general: Default::default(),
-                            audio: Default::default(),
-                            input: Default::default(),
-                            deepgram: None,
-                            groq: None,
-                            openai: None,
-                            local_whisper: None,
-                            local_vosk: None,
-                            local_parakeet: None,
-                            asr_sidecar: None,
-                            openai_compatible_realtime: None,
-                            llm: None,
-                            tts: None,
-                            hotkeys: None,
-                            llm_commands: Vec::new(),
-                            overlay: None,
-                        },
-                        Some(msg),
-                    );
-                }
-            },
-            Err(e) => {
-                let msg = format!(
-                    "Failed to read config at {}: {e} — using defaults",
-                    config_path.display()
-                );
-                error!("{msg}");
-                return (
-                    Config {
-                        general: Default::default(),
-                        audio: Default::default(),
-                        input: Default::default(),
-                        deepgram: None,
-                        groq: None,
-                        openai: None,
-                        local_whisper: None,
-                        local_vosk: None,
-                        local_parakeet: None,
-                        asr_sidecar: None,
-                        openai_compatible_realtime: None,
-                        llm: None,
-                        tts: None,
-                        hotkeys: None,
-                        llm_commands: Vec::new(),
-                        overlay: None,
-                    },
-                    Some(msg),
-                );
-            }
-        }
-    } else {
-        info!(
-            "no config file found at {}; using defaults",
-            config_path.display()
-        );
-    }
-    (
-        Config {
-            general: Default::default(),
-            audio: Default::default(),
-            input: Default::default(),
-            deepgram: None,
-            groq: None,
-            openai: None,
-            local_whisper: None,
-            local_vosk: None,
-            local_parakeet: None,
-            asr_sidecar: None,
-            openai_compatible_realtime: None,
-            llm: None,
-            tts: None,
-            hotkeys: None,
-            llm_commands: Vec::new(),
-            overlay: None,
-        },
-        None,
-    )
-}
-
-/// Maximum number of attempts to detect compositor environment.
-const COMPOSITOR_ENV_MAX_RETRIES: u32 = 10;
-
-/// Initial retry delay for compositor env detection (doubles each attempt, capped at 10 s).
-const COMPOSITOR_ENV_INITIAL_DELAY: Duration = Duration::from_secs(1);
-
-/// Minimum recording duration accepted by the gate. Anything shorter is almost
-/// certainly an accidental hotkey tap.
-const AUDIO_GATE_MIN_MS: u64 = 300;
-
-/// Compositor environment variables to import from systemd.
-const COMPOSITOR_ENV_VARS: &[&str] = &[
-    "WAYLAND_DISPLAY",
-    "DISPLAY",
-    "HYPRLAND_INSTANCE_SIGNATURE",
-    "SWAYSOCK",
-    "XDG_CURRENT_DESKTOP",
-];
-
-/// Wait for compositor environment variables to become available.
-///
-/// When the daemon starts via systemd on boot, it may launch before the
-/// compositor sets session environment variables (WAYLAND_DISPLAY, etc.).
-/// Without these, clipboard operations (wl-paste) and window tracking fail.
-///
-/// Polls `systemctl --user show-environment` with exponential backoff until
-/// a display server variable is found, then imports all compositor-related
-/// vars into the process environment.
-async fn import_compositor_env() {
-    // Already have a display server — nothing to do.
-    if std::env::var("WAYLAND_DISPLAY").is_ok() || std::env::var("DISPLAY").is_ok() {
-        debug!("compositor environment already available");
-        return;
-    }
-
-    info!("compositor env vars not set — polling systemd user environment");
-
-    let mut delay = COMPOSITOR_ENV_INITIAL_DELAY;
-
-    for attempt in 1..=COMPOSITOR_ENV_MAX_RETRIES {
-        if let Some(imported) = try_import_from_systemd() {
-            info!("imported compositor environment from systemd (attempt {attempt}): {imported}");
-            return;
-        }
-
-        if attempt == COMPOSITOR_ENV_MAX_RETRIES {
-            warn!(
-                "compositor environment not available after {COMPOSITOR_ENV_MAX_RETRIES} attempts \
-                 — clipboard and window tracking may not work"
-            );
-            return;
-        }
-
-        info!(
-            "compositor env not available (attempt {attempt}/{COMPOSITOR_ENV_MAX_RETRIES}) \
-             — retrying in {delay:?}"
-        );
-        tokio::time::sleep(delay).await;
-        delay = (delay * 2).min(Duration::from_secs(10));
-    }
-}
-
-/// Try to read compositor env vars from systemd's user environment.
-///
-/// Returns a summary string of imported vars on success, or None if no
-/// display server variable was found.
-fn try_import_from_systemd() -> Option<String> {
-    let output = std::process::Command::new("systemctl")
-        .args(["--user", "show-environment"])
-        .output()
-        .ok()?;
-
-    if !output.status.success() {
-        return None;
-    }
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    let mut imported = Vec::new();
-
-    for line in stdout.lines() {
-        if let Some((key, value)) = line.split_once('=') {
-            if COMPOSITOR_ENV_VARS.contains(&key) && std::env::var(key).is_err() {
-                std::env::set_var(key, value);
-                imported.push(key.to_string());
-            }
-        }
-    }
-
-    // Only succeed if we found a display server.
-    if std::env::var("WAYLAND_DISPLAY").is_ok() || std::env::var("DISPLAY").is_ok() {
-        Some(imported.join(", "))
-    } else {
-        None
-    }
-}
-
-fn check_uinput_access() {
-    use std::fs::OpenOptions;
-    match OpenOptions::new().write(true).open("/dev/uinput") {
-        Ok(_) => info!("uinput access: ok"),
-        Err(e) => {
-            if e.kind() == std::io::ErrorKind::PermissionDenied {
-                warn!(
-                    "Cannot open /dev/uinput — permission denied.\n\
-                     Fix: sudo usermod -aG input $USER\n\
-                          # Then log out and log back in\n\
-                     Or install the udev rule:\n\
-                          sudo install -m644 contrib/99-whisrs.rules /etc/udev/rules.d/\n\
-                          # On NixOS/Guix, point the rule at your setfacl:\n\
-                          command -v setfacl >/dev/null && sudo sed -i \\\n\
-                              \"s|/usr/bin/setfacl|$(command -v setfacl)|g\" \\\n\
-                              /etc/udev/rules.d/99-whisrs.rules\n\
-                          sudo udevadm control --reload-rules\n\
-                          sudo udevadm trigger"
-                );
-            } else {
-                warn!("Cannot open /dev/uinput: {e}");
-            }
-        }
-    }
-}
-
-fn check_audio_devices() {
-    use cpal::traits::{DeviceTrait, HostTrait};
-    let host = cpal::default_host();
-    match host.default_input_device() {
-        Some(device) => {
-            let name = device.name().unwrap_or_else(|_| "unknown".into());
-            info!("default audio input device: {name}");
-        }
-        None => {
-            warn!("no default audio input device found");
-            if let Ok(devices) = host.input_devices() {
-                let names: Vec<String> = devices.filter_map(|d| d.name().ok()).collect();
-                if names.is_empty() {
-                    warn!("no audio input devices available at all");
-                } else {
-                    warn!("available audio input devices: {}", names.join(", "));
-                }
-            }
-        }
-    }
-}
-
-fn validate_config(config: &Config) {
-    match config.validate() {
-        Ok(warnings) => {
-            for w in &warnings {
-                warn!("config: {}", w);
-            }
-        }
-        Err(e) => error!("config: {e}"),
-    }
-    if !config.has_any_backend_configured() {
-        warn!("No transcription backend configured. Run 'whisrs setup' to get started.");
-    }
-}
-
-fn resolve_groq_api_key(config: &Config) -> Option<String> {
-    if let Ok(key) = std::env::var("WHISRS_GROQ_API_KEY") {
-        if !key.is_empty() {
-            return Some(key);
-        }
-    }
-    config.groq.as_ref().map(|g| g.api_key.clone())
-}
-
-/// Resolve the API key used for TTS.
-///
-/// The dedicated `[tts] api_key` always wins. Otherwise the key is resolved
-/// per the configured `[tts] backend`, falling back to that provider's
-/// transcription key (env var or `config.toml`):
-/// - `groq`                       → Groq key (`[groq]` / `WHISRS_GROQ_API_KEY`)
-/// - `openai`                     → OpenAI key (`[openai]` / `WHISRS_OPENAI_API_KEY`)
-/// - `deepgram`                   → Deepgram key (`[deepgram]` / `WHISRS_DEEPGRAM_API_KEY`)
-/// - `tts-sidecar` / `openai-compat` → no key (local servers usually need none)
-fn resolve_tts_api_key(config: &Config) -> Option<String> {
-    if let Some(tts) = &config.tts {
-        if let Some(key) = tts.api_key.as_ref().filter(|k| !k.is_empty()) {
-            return Some(key.clone());
-        }
-        match tts.backend.as_str() {
-            "openai" => return resolve_openai_api_key(config),
-            "deepgram" => return resolve_deepgram_api_key(config),
-            "tts-sidecar" | "openai-compat" => return None,
-            _ => {}
-        }
-    }
-    resolve_groq_api_key(config)
-}
-
-fn resolve_openai_api_key(config: &Config) -> Option<String> {
-    if let Ok(key) = std::env::var("WHISRS_OPENAI_API_KEY") {
-        if !key.is_empty() {
-            return Some(key);
-        }
-    }
-    config.openai.as_ref().map(|o| o.api_key.clone())
-}
-
-fn resolve_deepgram_api_key(config: &Config) -> Option<String> {
-    if let Ok(key) = std::env::var("WHISRS_DEEPGRAM_API_KEY") {
-        if !key.is_empty() {
-            return Some(key);
-        }
-    }
-    config.deepgram.as_ref().map(|d| d.api_key.clone())
-}
-
-fn sanitize_ws_endpoint_for_log(url: &str) -> String {
-    let Ok(mut parsed) = reqwest::Url::parse(url) else {
-        return "<invalid ws endpoint>".to_string();
-    };
-    let _ = parsed.set_username("");
-    let _ = parsed.set_password(None);
-    parsed.set_query(None);
-    parsed.set_fragment(None);
-    parsed.to_string()
-}
-
-fn create_backend(config: &Config) -> Arc<dyn TranscriptionBackend> {
-    match config.general.backend.as_str() {
-        "deepgram" => {
-            let api_key = resolve_deepgram_api_key(config).unwrap_or_default();
-            if api_key.is_empty() {
-                warn!("no Deepgram API key configured");
-            }
-            info!("using Deepgram REST transcription backend");
-            Arc::new(DeepgramRestBackend::new(api_key))
-        }
-        "deepgram-streaming" => {
-            let api_key = resolve_deepgram_api_key(config).unwrap_or_default();
-            if api_key.is_empty() {
-                warn!("no Deepgram API key configured");
-            }
-            info!("using Deepgram streaming transcription backend");
-            Arc::new(DeepgramStreamingBackend::new(api_key))
-        }
-        "groq" => {
-            let api_key = resolve_groq_api_key(config).unwrap_or_default();
-            if api_key.is_empty() {
-                warn!("no Groq API key configured");
-            }
-            info!("using Groq transcription backend");
-            Arc::new(GroqBackend::new(api_key))
-        }
-        "openai-realtime" => {
-            let api_key = resolve_openai_api_key(config).unwrap_or_default();
-            if api_key.is_empty() {
-                warn!("no OpenAI API key configured");
-            }
-            info!("using OpenAI Realtime transcription backend");
-            Arc::new(OpenAIRealtimeBackend::new(api_key))
-        }
-        "openai" => {
-            let api_key = resolve_openai_api_key(config).unwrap_or_default();
-            if api_key.is_empty() {
-                warn!("no OpenAI API key configured");
-            }
-            info!("using OpenAI REST transcription backend");
-            Arc::new(OpenAIRestBackend::new(api_key))
-        }
-        "local-whisper" | "local" => {
-            // Serde defaults only apply when the [local-whisper] section
-            // exists; `LocalWhisperConfig::new` supplies the same defaults
-            // for a fully absent section.
-            let local_whisper = config.local_whisper.clone().unwrap_or_else(|| {
-                LocalWhisperConfig::new(
-                    dirs::data_dir()
-                        .unwrap_or_else(|| std::path::PathBuf::from("~/.local/share"))
-                        .join("whisrs/models/ggml-base.en.bin")
-                        .to_string_lossy()
-                        .to_string(),
-                )
-            });
-            info!(
-                "using local whisper transcription backend \
-                 (model: {}, segmentation: {})",
-                local_whisper.model_path, local_whisper.segmentation
-            );
-            Arc::new(
-                LocalWhisperBackend::new(local_whisper.model_path).with_segmentation(
-                    &local_whisper.segmentation,
-                    local_whisper.phrase_silence_ms,
-                ),
-            )
-        }
-        "local-vosk" => {
-            let model_path = config
-                .local_vosk
-                .as_ref()
-                .map(|l| l.model_path.clone())
-                .unwrap_or_else(|| {
-                    dirs::data_dir()
-                        .unwrap_or_else(|| std::path::PathBuf::from("~/.local/share"))
-                        .join("whisrs/models/vosk-model-small-en-us-0.15")
-                        .to_string_lossy()
-                        .to_string()
-                });
-            info!("using Vosk transcription backend (model: {model_path})");
-            Arc::new(VoskBackend::new(model_path))
-        }
-        "local-parakeet" => {
-            let model_path = config
-                .local_parakeet
-                .as_ref()
-                .map(|l| l.model_path.clone())
-                .unwrap_or_else(|| {
-                    dirs::data_dir()
-                        .unwrap_or_else(|| std::path::PathBuf::from("~/.local/share"))
-                        .join("whisrs/models/parakeet-eou-120m")
-                        .to_string_lossy()
-                        .to_string()
-                });
-            info!("using Parakeet transcription backend (model: {model_path})");
-            Arc::new(ParakeetBackend::new(model_path))
-        }
-        "asr-sidecar" | "asr" | "vibevoice" => {
-            let url = config
-                .asr_sidecar
-                .as_ref()
-                .map(|v| v.url.clone())
-                .unwrap_or_else(|| "http://127.0.0.1:8765/transcribe".to_string());
-            info!("using ASR sidecar transcription backend ({url})");
-            Arc::new(AsrSidecarBackend::new(url))
-        }
-        "openai-compatible-realtime" => {
-            let realtime = config.openai_compatible_realtime.as_ref().cloned();
-            let Some(realtime) = realtime else {
-                warn!(
-                    "openai-compatible-realtime backend selected but config section is missing; falling back to groq"
-                );
-                let api_key = resolve_groq_api_key(config).unwrap_or_default();
-                return Arc::new(GroqBackend::new(api_key));
-            };
-
-            let endpoint_display = sanitize_ws_endpoint_for_log(&realtime.url);
-            info!(
-                "using OpenAI-compatible realtime transcription backend ({endpoint_display}, profile={}, turn_detection={})",
-                realtime.profile, realtime.turn_detection
-            );
-
-            match OpenAiCompatibleRealtimeBackend::new(
-                realtime.url,
-                realtime.model,
-                realtime.profile,
-                realtime.turn_detection,
-                realtime.api_key,
-            ) {
-                Ok(backend) => Arc::new(backend),
-                Err(e) => {
-                    warn!(
-                        "failed to initialize OpenAI-compatible realtime backend for {endpoint_display}: {e}; falling back to groq"
-                    );
-                    let api_key = resolve_groq_api_key(config).unwrap_or_default();
-                    Arc::new(GroqBackend::new(api_key))
-                }
-            }
-        }
-        other => {
-            warn!("unknown backend '{other}', falling back to groq");
-            let api_key = resolve_groq_api_key(config).unwrap_or_default();
-            Arc::new(GroqBackend::new(api_key))
-        }
-    }
-}
-
-fn get_model_for_backend(config: &Config) -> String {
-    match config.general.backend.as_str() {
-        "deepgram" | "deepgram-streaming" => config
-            .deepgram
-            .as_ref()
-            .map(|d| d.model.clone())
-            .unwrap_or_else(|| "nova-3".to_string()),
-        "groq" => config
-            .groq
-            .as_ref()
-            .map(|g| g.model.clone())
-            .unwrap_or_else(|| "whisper-large-v3-turbo".to_string()),
-        "openai-realtime" => config
-            .openai
-            .as_ref()
-            .map(|o| o.model.clone())
-            .unwrap_or_else(|| "gpt-realtime-whisper".to_string()),
-        "openai" => config
-            .openai
-            .as_ref()
-            .map(|o| o.model.clone())
-            .unwrap_or_else(|| "gpt-4o-mini-transcribe".to_string()),
-        "openai-compatible-realtime" => config
-            .openai_compatible_realtime
-            .as_ref()
-            .map(|v| v.model.clone())
-            .unwrap_or_else(|| "Whisper-Tiny".to_string()),
-        "local-whisper" | "local" => "base.en".to_string(),
-        "local-vosk" => "small-en-us".to_string(),
-        "local-parakeet" => "eou-120m".to_string(),
-        "asr-sidecar" | "asr" | "vibevoice" => config
-            .asr_sidecar
-            .as_ref()
-            .map(|v| v.model.clone())
-            .unwrap_or_else(|| "microsoft/VibeVoice-ASR-HF".to_string()),
-        _ => "whisper-large-v3-turbo".to_string(),
-    }
-}
-
-/// Truncate a string to at most `max_bytes` bytes, respecting UTF-8 char
-/// boundaries. Appends "..." if truncated.
-fn truncate_preview(text: &str, max_bytes: usize) -> String {
-    if text.len() <= max_bytes {
-        return text.to_string();
-    }
-    let end = text
-        .char_indices()
-        .map(|(i, _)| i)
-        .take_while(|&i| i <= max_bytes)
-        .last()
-        .unwrap_or(0);
-    format!("{}...", &text[..end])
-}
-
-fn send_notification(summary: &str, body: &str) {
-    let summary = summary.to_string();
-    let body = body.to_string();
-    // Run on a blocking thread to avoid D-Bus block_on conflict with ksni tray.
-    std::thread::spawn(move || {
-        if let Err(e) = notify_rust::Notification::new()
-            .summary(&summary)
-            .body(&body)
-            .appname("whisrs")
-            .timeout(notify_rust::Timeout::Milliseconds(2000))
-            .show()
-        {
-            warn!("failed to send notification: {e}");
-        }
-    });
-}
 
 /// The daemon takes no options of its own, but declaring the interface means
 /// `--version` and `--help` are answered and exit, instead of being ignored and
@@ -918,2622 +241,6 @@ async fn handle_command(
     }
 }
 
-async fn handle_toggle(
-    daemon_state: Arc<Mutex<DaemonState>>,
-    context: Arc<DaemonContext>,
-    language: Option<String>,
-) -> Response {
-    let mut ds = daemon_state.lock().await;
-    let current_state = ds.state_machine.state();
-
-    match current_state {
-        State::Idle => {
-            // Fail fast on a bad `-l` override before any capture or state
-            // transition — otherwise the user records a whole session and
-            // only finds out when the backend rejects the language.
-            let language = match validate_toggle_language(language) {
-                Ok(lang) => lang,
-                Err(response) => return response,
-            };
-
-            // Resolve the session language once: per-toggle override, else
-            // config default. Persisted in `DaemonState` below so the
-            // stop-toggle can apply it to the batch path and history too.
-            let session_language = resolve_language(language, &context.config.general.language);
-
-            // Capture focused window before recording.
-            let window_id = match context.window_tracker.get_focused_window() {
-                Ok(id) => {
-                    info!("captured source window: {id}");
-                    Some(id)
-                }
-                Err(e) => {
-                    warn!("failed to capture focused window: {e}");
-                    None
-                }
-            };
-
-            // Start recording.
-            let mut capture =
-                match AudioCaptureHandle::start_with_level_tx(context.overlay_level_tx.clone()) {
-                    Ok(c) => c,
-                    Err(e) => {
-                        let msg = format!("{e}");
-                        let friendly = if msg.contains("no default audio input device") {
-                            format_no_microphone_error()
-                        } else {
-                            format!("Failed to start audio capture: {e}")
-                        };
-                        error!("{friendly}");
-                        return Response::Error { message: friendly };
-                    }
-                };
-
-            // For streaming backends: start the streaming pipeline immediately.
-            // Audio flows in real-time from microphone → API → text at cursor.
-            if context.transcription_backend.supports_streaming() {
-                let audio_rx = match capture.take_receiver() {
-                    Some(rx) => rx,
-                    None => {
-                        return Response::Error {
-                            message: "failed to get audio receiver".to_string(),
-                        }
-                    }
-                };
-
-                let config = build_transcription_config(&context.config, &session_language);
-
-                let backend = Arc::clone(&context.transcription_backend);
-                let wid = window_id.clone();
-                let ctx_notify = context.notify;
-                let ctx_overlay = context.overlay_enabled;
-                let window_tracker_for_pipeline = Arc::clone(&context.window_tracker);
-                // Restore focus before starting the pipeline.
-                let wid_for_focus = wid.clone();
-
-                // Spawn a task to:
-                // 1. Run auto-stop detection + forward audio to transcription
-                // 2. Run transcription backend
-                // 3. Type text as it arrives
-                let silence_timeout = context.config.general.silence_timeout_ms;
-                let ds_ref = Arc::clone(&daemon_state);
-                let filler_enabled = context.config.general.remove_filler_words;
-                let filler_words = context.config.general.filler_words.clone();
-                let pipeline_audio_feedback = context.config.general.audio_feedback;
-                let pipeline_audio_volume = context.config.general.audio_feedback_volume;
-
-                let pipeline_backend_name = context.config.general.backend.clone();
-                let pipeline_language = session_language.clone();
-                let pipeline_state_tx = context.state_tx.clone();
-                let pipeline_key_delay =
-                    std::time::Duration::from_millis(context.config.input.key_delay_ms);
-                let pipeline_injector_backend = context.config.input.backend;
-
-                // Per-recording cancel flag: lets `handle_cancel` stop the
-                // typing loop, which keeps running detached when the outer
-                // pipeline future is aborted.
-                let cancel_flag = Arc::new(AtomicBool::new(false));
-                let pipeline_cancel = Arc::clone(&cancel_flag);
-
-                let task = tokio::spawn(async move {
-                    run_streaming_pipeline(
-                        audio_rx,
-                        backend,
-                        config,
-                        wid,
-                        ctx_notify,
-                        ctx_overlay,
-                        silence_timeout,
-                        ds_ref,
-                        window_tracker_for_pipeline,
-                        filler_enabled,
-                        filler_words,
-                        pipeline_audio_feedback,
-                        pipeline_audio_volume,
-                        pipeline_backend_name,
-                        pipeline_language,
-                        pipeline_state_tx,
-                        pipeline_key_delay,
-                        pipeline_injector_backend,
-                        pipeline_cancel,
-                    )
-                    .await
-                });
-
-                ds.streaming_task = Some(task);
-                ds.streaming_cancel = Some(cancel_flag);
-
-                // Focus the window now (so text goes to the right place from the start).
-                if let Some(wid) = &wid_for_focus {
-                    if let Err(e) = context.window_tracker.focus_window(wid) {
-                        warn!("failed to pre-focus window: {e}");
-                    }
-                }
-            }
-
-            ds.audio_capture = Some(capture);
-            ds.recording_window_id = window_id;
-            ds.recording_started_at = Some(std::time::Instant::now());
-            ds.session_language = Some(session_language);
-
-            match ds.state_machine.transition(Action::Toggle) {
-                Ok(new_state) => {
-                    info!("started recording");
-                    // Broadcast recording state for tray.
-                    let _ = context.state_tx.send(new_state);
-                    if context.config.general.audio_feedback {
-                        feedback::play_start(context.config.general.audio_feedback_volume);
-                    }
-                    if context.notify_state() {
-                        send_notification("whisrs", "Recording...");
-                    }
-                    Response::Ok { state: new_state }
-                }
-                Err(e) => {
-                    ds.audio_capture = None;
-                    ds.recording_window_id = None;
-                    ds.streaming_task = None;
-                    ds.streaming_cancel = None;
-                    ds.session_language = None;
-                    Response::Error {
-                        message: e.to_string(),
-                    }
-                }
-            }
-        }
-        State::Recording => {
-            // Stop recording.
-            match ds.state_machine.transition(Action::Toggle) {
-                Ok(_) => {
-                    info!("stopped recording, transitioning to transcribing");
-                    // Broadcast transcribing state for tray.
-                    let _ = context.state_tx.send(State::Transcribing);
-                    if let Some(level_tx) = &context.overlay_level_tx {
-                        let _ = level_tx.send(0.0);
-                    }
-                    if context.config.general.audio_feedback {
-                        feedback::play_stop(context.config.general.audio_feedback_volume);
-                    }
-                    if context.notify_state() {
-                        send_notification("whisrs", "Transcribing...");
-                    }
-
-                    let capture = ds.audio_capture.take();
-                    let window_id = ds.recording_window_id.take();
-                    let streaming_task = ds.streaming_task.take();
-                    // Normal stop: drop the cancel flag untriggered so the
-                    // pipeline drains and types the remaining text.
-                    ds.streaming_cancel = None;
-                    let recording_started_at = ds.recording_started_at.take();
-                    // The language this session was started with. Consuming it
-                    // here ends the language session.
-                    let session_language =
-                        ds.take_session_language(&context.config.general.language);
-
-                    // A `-l` on the stop-press comes too late to change the
-                    // session — tell the user instead of silently dropping it.
-                    if let Some(requested) = &language {
-                        if *requested != session_language {
-                            warn!(
-                                "language override '{requested}' ignored — this session was \
-                                 started with '{session_language}'; pass -l on the toggle that \
-                                 starts recording"
-                            );
-                        }
-                    }
-
-                    // Release lock before slow operations.
-                    drop(ds);
-
-                    let result = if let Some(task) = streaming_task {
-                        // Streaming path: stop capture to close the channel,
-                        // then wait for the pipeline to drain and finish.
-                        if let Some(mut cap) = capture {
-                            cap.stop();
-                            tokio::task::spawn_blocking(move || drop(cap));
-                        }
-                        match task.await {
-                            Ok(Ok(text)) => Ok(text),
-                            Ok(Err(e)) => Err(e),
-                            Err(e) => Err(anyhow::anyhow!("streaming task panicked: {e}")),
-                        }
-                    } else {
-                        // Batch path: collect all audio, then transcribe with
-                        // the session language (not the config default).
-                        process_recording_batch(
-                            capture,
-                            window_id.as_deref(),
-                            &context,
-                            &session_language,
-                        )
-                        .await
-                    };
-
-                    // Transition back to Idle.
-                    let duration_secs = recording_started_at
-                        .map(|t| t.elapsed().as_secs_f64())
-                        .unwrap_or(0.0);
-                    let mut ds = daemon_state.lock().await;
-                    match ds.state_machine.transition(Action::TranscriptionDone) {
-                        Ok(new_state) => {
-                            // Broadcast idle state for tray.
-                            let _ = context.state_tx.send(new_state);
-                            if let Some(level_tx) = &context.overlay_level_tx {
-                                let _ = level_tx.send(0.0);
-                            }
-                            match result {
-                                Ok(text) => {
-                                    info!("transcription complete: {} chars", text.len());
-                                    if !text.is_empty() {
-                                        save_history_entry(
-                                            &text,
-                                            &context.config.general.backend,
-                                            &session_language,
-                                            duration_secs,
-                                        );
-                                    }
-                                    if context.config.general.audio_feedback {
-                                        feedback::play_done(
-                                            context.config.general.audio_feedback_volume,
-                                        );
-                                    }
-                                    if context.notify_state() {
-                                        let preview = truncate_preview(&text, 77);
-                                        send_notification("whisrs", &format!("Done: {preview}"));
-                                    }
-                                    Response::Ok { state: new_state }
-                                }
-                                Err(e) => {
-                                    error!("transcription failed: {e:#}");
-                                    if context.notify_error() {
-                                        send_notification(
-                                            "whisrs",
-                                            &format!("Transcription failed: {e}"),
-                                        );
-                                    }
-                                    Response::Ok { state: new_state }
-                                }
-                            }
-                        }
-                        // For STREAMING backends, run_streaming_pipeline's tail
-                        // often already moved Transcribing → Idle, so this
-                        // second TranscriptionDone is invalid. That is not an
-                        // error — the pipeline finalized everything (history,
-                        // done-tone, toast). Report success without double-
-                        // saving or double-notifying.
-                        Err(_) if ds.state_machine.state() == State::Idle => {
-                            let _ = context.state_tx.send(State::Idle);
-                            if let Some(level_tx) = &context.overlay_level_tx {
-                                let _ = level_tx.send(0.0);
-                            }
-                            match &result {
-                                Ok(text) => {
-                                    info!(
-                                        "transcription complete (pipeline-finalized): {} chars",
-                                        text.len()
-                                    )
-                                }
-                                Err(e) => error!("transcription failed: {e:#}"),
-                            }
-                            Response::Ok { state: State::Idle }
-                        }
-                        Err(e) => Response::Error {
-                            message: e.to_string(),
-                        },
-                    }
-                }
-                Err(e) => Response::Error {
-                    message: e.to_string(),
-                },
-            }
-        }
-        State::Transcribing => {
-            if let Some(requested) = &language {
-                warn!(
-                    "language override '{requested}' ignored — daemon is still transcribing \
-                     the previous session"
-                );
-            }
-            Response::Error {
-                message: "cannot toggle while transcribing".to_string(),
-            }
-        }
-        // Recording is refused while read-aloud is active.
-        State::Synthesizing | State::Speaking => Response::Error {
-            message: "cannot toggle while reading aloud — cancel read-aloud first".to_string(),
-        },
-    }
-}
-
-/// The streaming pipeline: reads audio in real-time, sends to API, types text.
-/// Also monitors for silence auto-stop.
-#[allow(clippy::too_many_arguments)]
-async fn run_streaming_pipeline(
-    mut audio_rx: tokio::sync::mpsc::UnboundedReceiver<Vec<i16>>,
-    backend: Arc<dyn TranscriptionBackend>,
-    config: TranscriptionConfig,
-    window_id: Option<String>,
-    notify: bool,
-    overlay_enabled: bool,
-    silence_timeout_ms: u64,
-    daemon_state: Arc<Mutex<DaemonState>>,
-    window_tracker: Arc<dyn WindowTracker>,
-    filler_enabled: bool,
-    filler_words: Vec<String>,
-    audio_feedback: bool,
-    audio_feedback_volume: f32,
-    backend_name: String,
-    language: String,
-    state_tx: tokio::sync::watch::Sender<State>,
-    key_delay: std::time::Duration,
-    injector_backend: InjectorBackend,
-    cancel_flag: Arc<AtomicBool>,
-) -> Result<String> {
-    // State-progress toasts are noise when the overlay is on.
-    let notify_state = notify && !overlay_enabled;
-    let notify_error = notify;
-    let pipeline_start = std::time::Instant::now();
-    let (audio_tx, backend_rx) = tokio::sync::mpsc::channel::<Vec<i16>>(256);
-    let (text_tx, text_rx) = tokio::sync::mpsc::channel::<String>(64);
-
-    // Build the filler filter once for the lifetime of this pipeline so the
-    // batch loop below isn't recompiling regexes on every typed delta.
-    let filler_filter = if filler_enabled {
-        Some(
-            FillerFilter::new(&filler_words)
-                .context("invalid custom filler word in configuration")?,
-        )
-    } else {
-        None
-    };
-
-    // Spawn the transcription backend.
-    let config_clone = config.clone();
-    let backend_task = tokio::spawn(async move {
-        backend
-            .transcribe_stream(backend_rx, text_tx, &config_clone)
-            .await
-    });
-
-    // Spawn a task that batches and types text as it arrives.
-    // We collect deltas for a short window to avoid creating a new virtual
-    // keyboard for every single word delta from the streaming API.
-    let wid = window_id.clone();
-    let typing_cancel = Arc::clone(&cancel_flag);
-    let typing_task = tokio::spawn(async move {
-        // Focus the original window before the first batch (only once).
-        // Sequenced by the batcher awaiting each sink call, so a plain
-        // atomic is enough to carry the flag into the 'static sink futures.
-        let focused = Arc::new(AtomicBool::new(false));
-        run_typing_batcher(text_rx, typing_cancel, filler_filter, move |text_to_type| {
-            let wid = wid.clone();
-            let tracker = Arc::clone(&window_tracker);
-            let focused = Arc::clone(&focused);
-            async move {
-                // Focus the original window (only once, or re-focus if needed).
-                if !focused.swap(true, Ordering::SeqCst) {
-                    if let Some(wid) = &wid {
-                        if let Err(e) = tracker.focus_window(wid) {
-                            warn!("failed to refocus window {wid} before typing: {e}");
-                        } else {
-                            tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-                        }
-                    }
-                }
-
-                info!("typing: {:?}", text_to_type);
-                // Streaming deliberately bypasses `inject_text` / `[input]
-                // paste`: partial deltas are typed as they arrive, and a
-                // paste per delta would thrash the clipboard.
-                match tokio::task::spawn_blocking(move || {
-                    type_text_at_cursor(&text_to_type, key_delay, injector_backend)
-                })
-                .await
-                {
-                    Ok(Ok(())) => {}
-                    Ok(Err(e)) => warn!("failed to type text: {e:#}"),
-                    Err(e) => warn!("failed to join typing task: {e}"),
-                }
-            }
-        })
-        .await
-    });
-
-    // Forward audio from capture to backend, with auto-stop detection.
-    let mut auto_stop =
-        AutoStopDetector::new(SILENCE_RMS_THRESHOLD, silence_timeout_ms, SAMPLE_RATE);
-
-    while let Some(chunk) = audio_rx.recv().await {
-        // Check for auto-stop.
-        if auto_stop.feed(&chunk) {
-            info!("silence auto-stop triggered after {silence_timeout_ms}ms");
-            if notify_state {
-                send_notification("whisrs", "Auto-stopped (silence detected)");
-            }
-
-            // Trigger stop: signal the daemon state machine.
-            // We stop the audio capture by closing the forwarding channel.
-            // The daemon state will be updated when the streaming task finishes.
-            let mut ds = daemon_state.lock().await;
-            if ds.state_machine.state() == State::Recording {
-                // Stop the audio capture.
-                if let Some(mut capture) = ds.audio_capture.take() {
-                    capture.stop();
-                    tokio::task::spawn_blocking(move || drop(capture));
-                }
-                // Transition to transcribing (pipeline is draining).
-                if let Err(e) = ds.state_machine.transition(Action::Toggle) {
-                    warn!("auto-stop state transition failed: {e}");
-                } else {
-                    let _ = state_tx.send(ds.state_machine.state());
-                }
-            }
-            break;
-        }
-
-        // Forward to backend.
-        if audio_tx.send(chunk).await.is_err() {
-            break;
-        }
-    }
-
-    // Drain remaining audio from the capture channel into the backend.
-    while let Some(chunk) = audio_rx.recv().await {
-        if audio_tx.send(chunk).await.is_err() {
-            break;
-        }
-    }
-
-    // Close the audio channel to signal end-of-stream to the backend.
-    drop(audio_tx);
-
-    // Wait for backend to finish.
-    let mut stream_error: Option<String> = None;
-    match backend_task.await {
-        Ok(Ok(())) => {}
-        Ok(Err(e)) => {
-            let friendly = format_api_error(&e);
-            error!("streaming transcription error: {friendly}");
-            stream_error = Some(friendly);
-        }
-        Err(e) => {
-            error!("streaming backend task panicked: {e}");
-            stream_error = Some(format!("transcription task panicked: {e}"));
-        }
-    }
-
-    // Wait for the typing side to observe the closed text channel and drain
-    // any final batch. If that somehow gets stuck, abort it so the daemon can
-    // return to Idle instead of staying in Transcribing forever.
-    debug!("waiting for typing task to finish");
-    let mut typing_task = typing_task;
-    let full_text = tokio::select! {
-        result = &mut typing_task => {
-            match result {
-                Ok(text) => text,
-                Err(e) => {
-                    warn!("typing task join failed during pipeline shutdown: {e}");
-                    String::new()
-                }
-            }
-        }
-        _ = tokio::time::sleep(TYPING_DRAIN_TIMEOUT) => {
-            warn!(
-                "typing task did not finish within {:?}; aborting to unblock daemon state",
-                TYPING_DRAIN_TIMEOUT
-            );
-            typing_task.abort();
-            match typing_task.await {
-                Ok(text) => text,
-                Err(e) if e.is_cancelled() => String::new(),
-                Err(e) => {
-                    warn!("typing task reported an unexpected shutdown error after abort: {e}");
-                    String::new()
-                }
-            }
-        }
-    };
-
-    // Notify user about streaming errors. Errors always pop, even with the
-    // overlay on — the overlay can't carry the failure detail.
-    if let Some(err_msg) = &stream_error {
-        if notify_error {
-            if full_text.is_empty() {
-                send_notification("whisrs", &format!("Transcription error: {err_msg}"));
-            } else {
-                send_notification(
-                    "whisrs",
-                    &format!("Transcription failed — partial text may have been typed.\n{err_msg}"),
-                );
-            }
-        }
-    }
-
-    // Save to history if we got any text.
-    if !full_text.is_empty() {
-        let duration_secs = pipeline_start.elapsed().as_secs_f64();
-        save_history_entry(&full_text, &backend_name, &language, duration_secs);
-    }
-
-    // If auto-stop happened, we need to transition to Idle.
-    let mut ds = daemon_state.lock().await;
-    if ds.state_machine.state() == State::Transcribing {
-        debug!("streaming pipeline transitioning daemon state back to idle");
-        ds.state_machine.transition(Action::TranscriptionDone).ok();
-        // Auto-stop ends the session without a stop-toggle — clear the
-        // session language so it can't leak into the next recording.
-        ds.session_language = None;
-        let _ = state_tx.send(ds.state_machine.state());
-        if audio_feedback {
-            feedback::play_done(audio_feedback_volume);
-        }
-        if notify_state {
-            let preview = truncate_preview(&full_text, 77);
-            if !preview.is_empty() {
-                send_notification("whisrs", &format!("Done: {preview}"));
-            }
-        }
-    }
-
-    Ok(full_text)
-}
-
-/// Batches streaming text deltas and hands them to `sink` (the key injector
-/// in production). This is the single choke point through which all streamed
-/// text reaches the cursor.
-///
-/// Deltas arriving within 150ms of each other are coalesced into one batch so
-/// we don't create a new virtual keyboard per word delta. Returns the full
-/// accumulated (typed) text. Exits when the text channel closes or `cancel`
-/// is set.
-///
-/// The cancel flag is checked when a delta arrives and again right before a
-/// batch reaches the sink, so a cancelled recording never types the text the
-/// backend flushes on shutdown (a batch already inside the sink still
-/// finishes). Exiting drops `text_rx`, which makes subsequent backend sends
-/// fail fast and winds the detached backend task down.
-///
-/// The daemon text channel is intentionally append-only. For OpenAI realtime
-/// this still feels token-by-token because the backend emits stable append
-/// deltas. For Lemonade-compatible profiles, the protocol layer only forwards
-/// completed utterances, so this loop naturally types phrase-sized chunks
-/// without needing any replacement semantics.
-async fn run_typing_batcher<F, Fut>(
-    mut text_rx: tokio::sync::mpsc::Receiver<String>,
-    cancel: Arc<AtomicBool>,
-    filler_filter: Option<FillerFilter>,
-    mut sink: F,
-) -> String
-where
-    F: FnMut(String) -> Fut,
-    Fut: std::future::Future<Output = ()>,
-{
-    let mut full_text = String::new();
-    let batch_delay = std::time::Duration::from_millis(150);
-
-    loop {
-        // Wait for the first delta (blocking).
-        let first = text_rx.recv().await;
-        let Some(first) = first else { break };
-
-        // Cancelled while we were waiting: discard everything from here on.
-        if cancel.load(Ordering::SeqCst) {
-            break;
-        }
-
-        // Collect this delta and any others that arrive within the batch window.
-        let mut batch = first;
-        while let Ok(Some(more)) = tokio::time::timeout(batch_delay, text_rx.recv()).await {
-            batch.push_str(&more);
-        }
-
-        if batch.is_empty() {
-            continue;
-        }
-
-        // Apply filler word removal if enabled.
-        if let Some(filter) = filler_filter.as_ref() {
-            batch = filter.apply(&batch);
-            if batch.is_empty() {
-                continue;
-            }
-        }
-
-        // Add space separator between turns if needed.
-        // Don't insert a space before punctuation streaming deltas,
-        // as they arrive as bare tokens.
-        let text_to_type = if full_text.is_empty()
-            || batch.starts_with(' ')
-            || full_text.ends_with(' ')
-            || leads_with_punct(&batch)
-        {
-            batch
-        } else {
-            format!(" {batch}")
-        };
-
-        // Last gate before the sink: cancel may have arrived while this
-        // batch was being collected.
-        if cancel.load(Ordering::SeqCst) {
-            break;
-        }
-
-        full_text.push_str(&text_to_type);
-        sink(text_to_type).await;
-    }
-
-    full_text
-}
-
-/// Where a batch transcription request originated. Selects the per-path
-/// notification/log wording inside [`transcribe_batch_audio`]; the behavior
-/// toggles live in [`BatchOptions`].
-enum BatchOrigin<'a> {
-    /// Plain dictation (`whisrs toggle` with a batch backend).
-    Dictation,
-    /// Command mode (`whisrs command`): selection + voice instruction.
-    CommandMode,
-    /// A named `[[llm_commands]]` hotkey. Carries the command name for
-    /// toasts/logs — runtime config data, hence not a `&'static str` label.
-    LlmCommand { name: &'a str },
-}
-
-/// Per-path switches for [`transcribe_batch_audio`].
-///
-/// File-local for now; the daemon module split moves this out later.
-struct BatchOptions<'a> {
-    /// Send the configured prompt/vocabulary hint via
-    /// `build_transcription_config` instead of a bare config.
-    use_prompt: bool,
-    /// Save the recording for `whisrs transcribe-recovery` when the backend
-    /// errors. Dictation only: command-path audio is an instruction, not
-    /// content worth replaying later.
-    save_recovery: bool,
-    /// Drop transcripts that look like the model echoing the prompt back.
-    /// Should be `true` whenever `use_prompt` is `true` — an echoed prompt is
-    /// garbage whether it would be typed or fed to the LLM.
-    check_prompt_echo: bool,
-    /// Apply the configured filler-word filter. Dictation only for now.
-    apply_filler: bool,
-    /// Which flow is asking; picks the notification/log wording.
-    origin: BatchOrigin<'a>,
-}
-
-/// Shared batch transcription flow: audio gate → WAV encode → backend
-/// transcribe → prompt-echo check → filler removal.
-///
-/// `Ok(String::new())` means "nothing to act on" — the recording was gated,
-/// the transcript was dropped as a prompt echo, or it came back empty — and
-/// the relevant user-facing toast has already been sent with the calling
-/// path's own wording (selected via `opts.origin`). `Err` is a real failure
-/// (encode or backend); the caller formats and reports it, then unwinds.
-async fn transcribe_batch_audio(
-    samples: &[i16],
-    context: &DaemonContext,
-    language: &str,
-    opts: &BatchOptions<'_>,
-) -> Result<String> {
-    use whisrs::audio::wav::encode_wav;
-
-    // Skip the API call entirely when the recording is empty, too short, or
-    // pure silence. Cloud Whisper variants (whisper-1, gpt-4o-*-transcribe)
-    // hallucinate verbatim chunks of the supplied prompt when handed audio
-    // with no speech, which whisrs would then type at the cursor — for a
-    // multi-hundred-character prompt that means tens of seconds of garbage
-    // typing on every accidental hotkey tap. Filtering here also saves the
-    // round-trip cost.
-    if let Some(reason) = audio_gate_reason(
-        samples,
-        SAMPLE_RATE,
-        AUDIO_GATE_MIN_MS,
-        SILENCE_RMS_THRESHOLD,
-    ) {
-        let label = reason.as_str();
-        match &opts.origin {
-            BatchOrigin::Dictation => {
-                info!(
-                    "skipping transcription: recording was {label} ({} samples)",
-                    samples.len()
-                );
-                if context.notify_state() {
-                    send_notification("whisrs", &format!("Skipped: recording was {label}"));
-                }
-            }
-            BatchOrigin::CommandMode => {
-                info!(
-                    "command mode: skipping (recording was {label}, {} samples)",
-                    samples.len()
-                );
-                if context.notify_error() {
-                    send_notification("whisrs", &format!("Command mode: recording was {label}"));
-                }
-            }
-            BatchOrigin::LlmCommand { name } => {
-                info!(
-                    "llm-command '{name}': skipping (recording was {label}, {} samples)",
-                    samples.len()
-                );
-                if context.notify_state() {
-                    send_notification("whisrs", &format!("Skipped: recording was {label}"));
-                }
-            }
-        }
-        return Ok(String::new());
-    }
-
-    // Progress toast for the command flows — after the gate, so a silent
-    // recording only surfaces the skip toast above.
-    match &opts.origin {
-        BatchOrigin::Dictation => {}
-        BatchOrigin::CommandMode => {
-            if context.notify_state() {
-                send_notification("whisrs", "Processing command...");
-            }
-        }
-        BatchOrigin::LlmCommand { name } => {
-            if context.notify_state() {
-                send_notification("whisrs", &format!("Processing '{name}'..."));
-            }
-        }
-    }
-
-    let wav_data = encode_wav(samples)?;
-    info!("encoded WAV: {} bytes", wav_data.len());
-
-    let config = if opts.use_prompt {
-        build_transcription_config(&context.config, language)
-    } else {
-        TranscriptionConfig {
-            language: language.to_string(),
-            model: get_model_for_backend(&context.config),
-            prompt: None,
-        }
-    };
-
-    let text = match context
-        .transcription_backend
-        .transcribe(&wav_data, &config)
-        .await
-    {
-        Ok(t) => t,
-        Err(e) => {
-            if opts.save_recovery {
-                let friendly = format_api_error(&e);
-                error!("transcription failed: {friendly}");
-                // Save audio for recovery.
-                use whisrs::audio::recovery;
-                match recovery::save_recovery_audio(samples) {
-                    Ok(path) => {
-                        info!(
-                            "audio saved for recovery: {} — retry with: whisrs transcribe-recovery",
-                            path.display()
-                        );
-                        if context.notify_error() {
-                            send_notification(
-                                "whisrs",
-                                &format!(
-                                    "Transcription failed: {friendly}\nAudio saved to {}\nRetry with: whisrs transcribe-recovery",
-                                    path.display()
-                                ),
-                            );
-                        }
-                    }
-                    Err(re) => {
-                        warn!("failed to save recovery audio: {re}");
-                    }
-                }
-            }
-            return Err(e);
-        }
-    };
-
-    // Defence in depth against prompt-echo hallucinations: even with the
-    // upstream gate, low-SNR recordings (tap microphones, very brief speech
-    // followed by silence) sometimes squeak through and trigger the model to
-    // regurgitate the prompt. Drop those before they reach the keyboard (or,
-    // for llm-commands, the LLM).
-    if opts.check_prompt_echo
-        && config
-            .prompt
-            .as_deref()
-            .is_some_and(|prompt| is_prompt_echo(&text, prompt))
-    {
-        warn!(
-            "discarding likely prompt-echo response ({} chars) — see prompt_echo crate",
-            text.len()
-        );
-        if context.notify_state() {
-            send_notification(
-                "whisrs",
-                "Skipped: response looked like a prompt echo (no speech detected)",
-            );
-        }
-        return Ok(String::new());
-    }
-
-    // Apply filler word removal if enabled.
-    let text = if opts.apply_filler && context.config.general.remove_filler_words {
-        let filter = FillerFilter::new(&context.config.general.filler_words)
-            .context("invalid custom filler word in configuration")?;
-        let cleaned = filter.apply(&text);
-        if cleaned != text {
-            info!(
-                "filler removal: {} chars -> {} chars",
-                text.len(),
-                cleaned.len()
-            );
-        }
-        cleaned
-    } else {
-        text
-    };
-
-    if text.is_empty() {
-        match &opts.origin {
-            BatchOrigin::Dictation => {
-                info!("transcription returned empty text — nothing to type");
-            }
-            BatchOrigin::CommandMode => {
-                if context.notify_error() {
-                    send_notification("whisrs", "Could not understand instruction — try again");
-                }
-            }
-            BatchOrigin::LlmCommand { .. } => {
-                if context.notify_error() {
-                    send_notification("whisrs", "Could not understand speech — try again");
-                }
-            }
-        }
-    }
-
-    Ok(text)
-}
-
-/// Batch mode: collect all audio, transcribe in one shot, type result.
-/// `language` is the resolved session language (per-toggle override or
-/// config default) captured when recording started.
-async fn process_recording_batch(
-    capture: Option<AudioCaptureHandle>,
-    window_id: Option<&str>,
-    context: &DaemonContext,
-    language: &str,
-) -> Result<String> {
-    let samples = match capture {
-        Some(cap) => cap.stop_and_collect().await?,
-        None => anyhow::bail!("no audio capture to collect"),
-    };
-
-    if samples.is_empty() {
-        anyhow::bail!("no audio samples captured");
-    }
-
-    info!("collected {} audio samples", samples.len());
-
-    let text = transcribe_batch_audio(
-        &samples,
-        context,
-        language,
-        &BatchOptions {
-            use_prompt: true,
-            save_recovery: true,
-            check_prompt_echo: true,
-            apply_filler: true,
-            origin: BatchOrigin::Dictation,
-        },
-    )
-    .await?;
-
-    if text.is_empty() {
-        // Gated, echo-dropped, or empty — the helper already said so.
-        return Ok(text);
-    }
-
-    // Restore window focus.
-    if let Some(wid) = window_id {
-        if let Err(e) = context.window_tracker.focus_window(wid) {
-            warn!("failed to restore window focus: {e}");
-        } else {
-            tokio::time::sleep(std::time::Duration::from_millis(100)).await;
-        }
-    }
-
-    // Inject the text at the cursor — type keystrokes, or paste via the
-    // clipboard when `[input] paste = true` (layout-independent).
-    let text_clone = text.clone();
-    let key_delay = std::time::Duration::from_millis(context.config.input.key_delay_ms);
-    let injector_backend = context.config.input.backend;
-    let paste = context.config.input.paste;
-    let is_terminal = if paste {
-        context
-            .window_tracker
-            .get_focused_window_class()
-            .map(|c| is_terminal_class(&c))
-            .unwrap_or(false)
-    } else {
-        false
-    };
-    match tokio::task::spawn_blocking(move || {
-        inject_text(&text_clone, is_terminal, key_delay, injector_backend, paste)
-    })
-    .await
-    {
-        Ok(Ok(())) => {}
-        Ok(Err(e)) => warn!("failed to inject text: {e:#}"),
-        Err(e) => warn!("failed to join injection task: {e}"),
-    }
-
-    Ok(text)
-}
-
-fn format_no_microphone_error() -> String {
-    use cpal::traits::{DeviceTrait, HostTrait};
-    let host = cpal::default_host();
-    let mut msg = "No microphone found — no default audio input device available.".to_string();
-    if let Ok(devices) = host.input_devices() {
-        let names: Vec<String> = devices.filter_map(|d| d.name().ok()).collect();
-        if names.is_empty() {
-            msg.push_str("\nNo audio input devices detected. Check that your microphone is connected and PipeWire/PulseAudio is running.");
-        } else {
-            msg.push_str("\nAvailable input devices:");
-            for name in &names {
-                msg.push_str(&format!("\n  - {name}"));
-            }
-            msg.push_str(
-                "\nSet the device in ~/.config/whisrs/config.toml under [audio] device = \"...\"",
-            );
-        }
-    }
-    msg
-}
-
-fn format_api_error(err: &anyhow::Error) -> String {
-    let msg = format!("{err}");
-    if msg.contains("final transcription completion") {
-        return "Realtime transcription timed out waiting for the server to finish after stop"
-            .to_string();
-    }
-    if msg.contains("invalid API key") || msg.contains("401") {
-        return "Invalid API key — check your config at ~/.config/whisrs/config.toml".to_string();
-    }
-    if msg.contains("rate limit") || msg.contains("429") {
-        return "Rate limited — wait a moment and try again".to_string();
-    }
-    if msg.contains("error sending request")
-        || msg.contains("dns error")
-        || msg.contains("connection refused")
-        || msg.contains("timed out")
-        || msg.contains("ConnectError")
-    {
-        return "Cannot reach API — check your internet connection".to_string();
-    }
-    msg
-}
-
-/// Type text at the cursor using uinput (keyboard injection) or clipboard paste.
-fn type_text_at_cursor(
-    text: &str,
-    key_delay: std::time::Duration,
-    backend: InjectorBackend,
-) -> Result<()> {
-    let keyboard_slot = KEYBOARD.get_or_init(|| StdMutex::new(None));
-    let mut keyboard_guard = keyboard_slot
-        .lock()
-        .map_err(|_| anyhow::anyhow!("keyboard mutex poisoned"))?;
-
-    if keyboard_guard.is_none() {
-        // Runtime path: short settle delay. The startup prewarm in
-        // `warm_keyboard` is what gives X11 time to attach its keymap;
-        // we don't want every error-recovery to stall the user's typing
-        // path for 1s.
-        *keyboard_guard = Some(new_keyboard(
-            key_delay, /* prewarm = */ false, backend,
-        )?);
-    }
-
-    let keyboard = keyboard_guard
-        .as_mut()
-        .expect("keyboard exists after initialization");
-    keyboard.set_key_delay(key_delay);
-
-    let result = keyboard.type_text(text).context("failed to type text");
-    if result.is_err() {
-        *keyboard_guard = None;
-    }
-    result?;
-    Ok(())
-}
-
-/// Send a paste keystroke — Ctrl+V, or Ctrl+Shift+V in terminals — via the
-/// **persistent** virtual keyboard (the same device `type_text_at_cursor`
-/// uses).
-///
-/// Must NOT use a fresh per-call uinput device: on some compositors (e.g.
-/// KWin) keystrokes from a device the compositor hasn't finished enumerating
-/// are dropped, so the paste silently no-ops. The persistent device is already
-/// recognized, so its keystrokes land. The combo is raw keycodes (`KEY_V` is
-/// `v` in every common layout), so it stays layout-independent.
-fn paste_via_keyboard(
-    is_terminal: bool,
-    key_delay: std::time::Duration,
-    backend: InjectorBackend,
-) -> Result<()> {
-    use evdev::Key;
-
-    let keyboard_slot = KEYBOARD.get_or_init(|| StdMutex::new(None));
-    let mut keyboard_guard = keyboard_slot
-        .lock()
-        .map_err(|_| anyhow::anyhow!("keyboard mutex poisoned"))?;
-
-    if keyboard_guard.is_none() {
-        *keyboard_guard = Some(new_keyboard(
-            key_delay, /* prewarm = */ false, backend,
-        )?);
-    }
-
-    let keyboard = keyboard_guard
-        .as_mut()
-        .expect("keyboard exists after initialization");
-    keyboard.set_key_delay(key_delay);
-
-    let combo: &[Key] = if is_terminal {
-        &[Key::KEY_LEFTCTRL, Key::KEY_LEFTSHIFT, Key::KEY_V]
-    } else {
-        &[Key::KEY_LEFTCTRL, Key::KEY_V]
-    };
-
-    let result = keyboard
-        .send_combo(combo)
-        .context("failed to send paste combo");
-    if result.is_err() {
-        *keyboard_guard = None;
-    }
-    result
-}
-
-/// Clear the current shell prompt line by sending Ctrl+A ("move to start of
-/// line") then Ctrl+K ("kill to end of line") via the **persistent** virtual
-/// keyboard (the same device `type_text_at_cursor` and `paste_via_keyboard`
-/// use). That readline / zle / fish editing pair empties the line in bash, zsh
-/// and fish alike.
-///
-/// Only ever called for terminals. A terminal's mouse highlight is a visual
-/// overlay, not an editable selection, so injecting at the cursor would append
-/// to the existing line rather than replace it. Clearing the line first makes
-/// the injected text the whole line, which is what a "rewrite my selection"
-/// command means at a prompt. In GUI text widgets no clear is needed (or
-/// wanted): typing/pasting over a real selection replaces it natively.
-///
-/// Must NOT use a fresh per-call uinput device: on some compositors (e.g.
-/// KWin) keystrokes from a device the compositor hasn't finished enumerating
-/// are dropped, so the clear silently no-ops. The persistent device is already
-/// recognized, so its keystrokes land.
-fn clear_line_via_keyboard(key_delay: std::time::Duration, backend: InjectorBackend) -> Result<()> {
-    use evdev::Key;
-
-    let keyboard_slot = KEYBOARD.get_or_init(|| StdMutex::new(None));
-    let mut keyboard_guard = keyboard_slot
-        .lock()
-        .map_err(|_| anyhow::anyhow!("keyboard mutex poisoned"))?;
-
-    if keyboard_guard.is_none() {
-        *keyboard_guard = Some(new_keyboard(
-            key_delay, /* prewarm = */ false, backend,
-        )?);
-    }
-
-    let keyboard = keyboard_guard
-        .as_mut()
-        .expect("keyboard exists after initialization");
-    keyboard.set_key_delay(key_delay);
-
-    let result = keyboard
-        .send_combo(&[Key::KEY_LEFTCTRL, Key::KEY_A])
-        .context("failed to send Ctrl+A (move to start of line)")
-        .and_then(|()| {
-            keyboard
-                .send_combo(&[Key::KEY_LEFTCTRL, Key::KEY_K])
-                .context("failed to send Ctrl+K (kill to end of line)")
-        });
-    if result.is_err() {
-        *keyboard_guard = None;
-    }
-    result
-}
-
-/// Inject `text` at the cursor, choosing keystrokes or clipboard paste.
-///
-/// With `paste = false` (default) this types via the virtual keyboard. With
-/// `paste = true` it sets the clipboard, sends Ctrl+V (Ctrl+Shift+V for
-/// terminals), then restores the previous clipboard — layout-independent
-/// injection for compositors that lack the Wayland virtual-keyboard protocol
-/// (see [`whisrs::InputConfig::paste`]). Runs in a blocking context (callers
-/// wrap it in `spawn_blocking`), so the sleeps/restore use std threads.
-///
-/// `ClipboardBackend` is text-only, so a clipboard holding non-text content
-/// (an image, a file list, ...) can't be captured and round-tripped at all.
-/// If the pre-paste read fails, this falls back to typing instead of pasting
-/// — overwriting the clipboard via `set_text` first and only then discovering
-/// there's nothing valid to restore would destroy that content permanently
-/// (see #69), which skipping the *restore* alone can't undo since the damage
-/// already happened at `set_text`.
-///
-/// The restore is otherwise skipped, rather than clobbering the clipboard,
-/// when the clipboard no longer holds the text we set — something else (the
-/// user, another app) copied over it during the paste, and restoring would
-/// race that copy and discard it.
-fn inject_text(
-    text: &str,
-    is_terminal: bool,
-    key_delay: std::time::Duration,
-    backend: InjectorBackend,
-    paste: bool,
-) -> Result<()> {
-    if !paste {
-        return type_text_at_cursor(text, key_delay, backend);
-    }
-
-    let clipboard = xkb_type::default_clipboard();
-    let saved = match clipboard.get_text() {
-        Ok(s) => s,
-        Err(e) => {
-            debug!("clipboard unreadable as text, typing instead of pasting: {e:#}");
-            return type_text_at_cursor(text, key_delay, backend);
-        }
-    };
-    clipboard
-        .set_text(text)
-        .context("failed to set clipboard for paste injection")?;
-
-    // Let the clipboard settle before the paste keystroke.
-    std::thread::sleep(std::time::Duration::from_millis(50));
-
-    let paste_result = paste_via_keyboard(is_terminal, key_delay, backend);
-
-    // Restore the user's clipboard after the paste has landed, regardless of
-    // whether the keystroke succeeded — but only if it's safe to (see doc
-    // comment above).
-    let pasted_text = text.to_string();
-    std::thread::spawn(move || {
-        std::thread::sleep(std::time::Duration::from_millis(500));
-        let clipboard = xkb_type::default_clipboard();
-        match clipboard.get_text() {
-            Ok(current) if current == pasted_text => {
-                if let Err(e) = clipboard.set_text(&saved) {
-                    warn!("failed to restore clipboard: {e}");
-                }
-            }
-            Ok(_) => {
-                debug!("clipboard changed during paste injection; skipping restore");
-            }
-            Err(e) => {
-                warn!("failed to read clipboard before restore, skipping restore: {e}");
-            }
-        }
-    });
-
-    paste_result
-}
-
-fn warm_keyboard(key_delay: std::time::Duration, backend: InjectorBackend) {
-    let keyboard_slot = KEYBOARD.get_or_init(|| StdMutex::new(None));
-    let Ok(mut keyboard_guard) = keyboard_slot.lock() else {
-        warn!("failed to initialize virtual keyboard: keyboard mutex poisoned");
-        return;
-    };
-
-    if keyboard_guard.is_some() {
-        return;
-    }
-
-    // Startup path: long settle delay so X11 has time to process
-    // MappingNotify and attach the device keymap before the first key.
-    match new_keyboard(key_delay, /* prewarm = */ true, backend) {
-        Ok(kb) => {
-            *keyboard_guard = Some(kb);
-            info!("virtual keyboard initialized");
-        }
-        Err(e) => {
-            warn!("failed to initialize virtual keyboard: {e:#}");
-        }
-    }
-}
-
-/// Build the uinput (evdev) keyboard, mapping the common permission failure
-/// to an actionable message. `prewarm` adds a settle delay so X11 attaches
-/// the device keymap before the first keystroke.
-fn new_uinput_keyboard(
-    key_delay: std::time::Duration,
-    prewarm: bool,
-) -> Result<Box<dyn xkb_type::KeyInjector>> {
-    let result = if prewarm {
-        xkb_type::Keyboard::new_prewarm(key_delay)
-    } else {
-        xkb_type::Keyboard::new(key_delay)
-    };
-    match result {
-        Ok(kb) => Ok(Box::new(kb)),
-        Err(e) => {
-            let msg = format!("{e:#}");
-            if msg.contains("Permission denied") || msg.contains("permission") {
-                anyhow::bail!(
-                    "Cannot open /dev/uinput — permission denied.\n\
-                     Fix: sudo usermod -aG input $USER"
-                );
-            }
-            Err(e.context("failed to create virtual keyboard"))
-        }
-    }
-}
-
-/// Construct the configured keyboard-injection backend.
-///
-/// `Auto` prefers the layout-independent Wayland virtual keyboard when a
-/// Wayland session is detected, falling back to uinput when the compositor
-/// lacks `zwp_virtual_keyboard_v1`. `prewarm` only affects the uinput path
-/// (the Wayland backend ships its own keymap, so no settle delay is needed).
-fn new_keyboard(
-    key_delay: std::time::Duration,
-    prewarm: bool,
-    backend: InjectorBackend,
-) -> Result<Box<dyn xkb_type::KeyInjector>> {
-    match backend {
-        InjectorBackend::Uinput => new_uinput_keyboard(key_delay, prewarm),
-        InjectorBackend::WaylandVk => {
-            let kb = xkb_type::wayland_vk::WaylandVkKeyboard::new(key_delay)?;
-            info!("using wayland virtual-keyboard injection backend");
-            Ok(Box::new(kb))
-        }
-        InjectorBackend::Auto => {
-            if std::env::var_os("WAYLAND_DISPLAY").is_some() {
-                match xkb_type::wayland_vk::WaylandVkKeyboard::new(key_delay) {
-                    Ok(kb) => {
-                        info!("using wayland virtual-keyboard injection backend");
-                        return Ok(Box::new(kb));
-                    }
-                    Err(e) => {
-                        warn!(
-                            "wayland virtual-keyboard unavailable, falling back to uinput: {e:#}"
-                        );
-                    }
-                }
-            }
-            new_uinput_keyboard(key_delay, prewarm)
-        }
-    }
-}
-
-/// Resolve the effective session language: the per-toggle `override_lang` when
-/// present, otherwise the configured default.
-fn resolve_language(override_lang: Option<String>, default_lang: &str) -> String {
-    override_lang.unwrap_or_else(|| default_lang.to_string())
-}
-
-/// Validate a per-toggle language override before recording starts.
-///
-/// Returns the normalized override on success, or the error `Response` to
-/// send back to the CLI. Called on the start-press before any capture or
-/// state transition so a bad `-l` value (e.g. `english`) is rejected up
-/// front instead of failing a whole session at transcription time. The
-/// config default (`general.language`) is deliberately not validated.
-fn validate_toggle_language(language: Option<String>) -> Result<Option<String>, Response> {
-    match language {
-        None => Ok(None),
-        Some(lang) => match validate_language_override(&lang) {
-            Ok(normalized) => Ok(Some(normalized)),
-            Err(message) => {
-                warn!("rejected language override: {message}");
-                Err(Response::Error { message })
-            }
-        },
-    }
-}
-
-fn build_transcription_config(config: &Config, language: &str) -> TranscriptionConfig {
-    TranscriptionConfig {
-        language: language.to_string(),
-        model: get_model_for_backend(config),
-        prompt: transcription_prompt(config.general.prompt.as_deref(), &config.general.vocabulary),
-    }
-}
-
-/// Joins `prompt` and `vocabulary` with `". "`, prompt first. Blank prompts
-/// are treated as absent; returns `None` only when both inputs are empty so
-/// backends skip the hint entirely rather than receiving an empty string.
-fn transcription_prompt(prompt: Option<&str>, vocabulary: &[String]) -> Option<String> {
-    let prompt = prompt.map(str::trim).filter(|s| !s.is_empty());
-    let vocab = if vocabulary.is_empty() {
-        None
-    } else {
-        Some(vocabulary.join(", "))
-    };
-    match (prompt, vocab) {
-        (Some(p), Some(v)) => Some(format!("{p}. {v}")),
-        (Some(p), None) => Some(p.to_string()),
-        (None, Some(v)) => Some(v),
-        (None, None) => None,
-    }
-}
-
-/// Save a transcription to the history file.
-fn save_history_entry(text: &str, backend: &str, language: &str, duration_secs: f64) {
-    let entry = HistoryEntry {
-        timestamp: chrono::Local::now(),
-        text: text.to_string(),
-        backend: backend.to_string(),
-        language: language.to_string(),
-        duration_secs,
-    };
-    if let Err(e) = history::append_entry(&entry) {
-        warn!("failed to save history entry: {e}");
-    }
-}
-
-/// Command mode toggle: first call copies selection and starts recording,
-/// second call stops recording and kicks off transcription → LLM → inject.
-/// Also auto-stops on silence.
-async fn handle_command_mode(
-    daemon_state: Arc<Mutex<DaemonState>>,
-    context: Arc<DaemonContext>,
-) -> Response {
-    let current_state = {
-        let ds = daemon_state.lock().await;
-        ds.state_machine.state()
-    };
-
-    match current_state {
-        State::Recording => {
-            // Second press: check if we're in command mode recording.
-            let is_command_mode = {
-                let ds = daemon_state.lock().await;
-                ds.command_mode.is_some()
-            };
-            if !is_command_mode {
-                return Response::Error {
-                    message: "recording is active but not in command mode — use toggle or cancel"
-                        .to_string(),
-                };
-            }
-            // Stop recording — the background task will detect the channel close.
-            let mut ds = daemon_state.lock().await;
-            if let Some(mut capture) = ds.audio_capture.take() {
-                capture.stop();
-                tokio::task::spawn_blocking(move || drop(capture));
-            }
-            info!("command mode: manual stop");
-            Response::Ok {
-                state: State::Recording,
-            }
-        }
-        State::Idle => {
-            // First press: copy selection and start recording.
-            command_mode_start(daemon_state, context).await
-        }
-        State::Transcribing => Response::Error {
-            message: "cannot start command mode while transcribing".to_string(),
-        },
-        // Command mode is a recording flow — refused while read-aloud is active.
-        State::Synthesizing | State::Speaking => Response::Error {
-            message: "cannot start command mode while reading aloud — cancel read-aloud first"
-                .to_string(),
-        },
-    }
-}
-
-/// Command mode first press: copy selection, start recording, spawn background processor.
-async fn command_mode_start(
-    daemon_state: Arc<Mutex<DaemonState>>,
-    context: Arc<DaemonContext>,
-) -> Response {
-    // Get LLM config.
-    let llm_config = context.config.llm.clone().unwrap_or_default();
-
-    // Step 1: Capture the selected text. Command mode prefers the primary
-    // selection (the X highlight, distinct from the Ctrl+C clipboard), which
-    // needs no key simulation and leaves the clipboard untouched. When the
-    // primary selection is empty it falls back to a simulated Ctrl+C, sharing
-    // the same capture path as read-aloud, so command mode still works on apps
-    // and compositors that don't populate the primary selection. The LLM result
-    // is later injected through the same wrapper dictation uses: typed by
-    // default, pasted when `[input] paste` is set. Either way it replaces the
-    // active selection in GUI apps and lands at the prompt cursor in terminals.
-    info!("command mode: getting selected text");
-    let selected_text = match capture_selection(&context).await {
-        Ok(text) => text,
-        Err(message) => return Response::Error { message },
-    };
-
-    info!(
-        "command mode: got {} chars of selected text",
-        selected_text.len()
-    );
-
-    // Step 2: Start recording voice instruction.
-    if context.config.general.audio_feedback {
-        feedback::play_start(context.config.general.audio_feedback_volume);
-    }
-
-    let mut capture =
-        match AudioCaptureHandle::start_with_level_tx(context.overlay_level_tx.clone()) {
-            Ok(c) => c,
-            Err(e) => {
-                return Response::Error {
-                    message: format!("failed to start audio capture: {e}"),
-                };
-            }
-        };
-
-    let audio_rx = capture.take_receiver();
-
-    // Store state.
-    {
-        let mut ds = daemon_state.lock().await;
-        if let Err(e) = ds.state_machine.transition(Action::Toggle) {
-            return Response::Error {
-                message: format!("state transition failed: {e}"),
-            };
-        }
-        ds.audio_capture = Some(capture);
-        ds.recording_started_at = Some(std::time::Instant::now());
-        ds.command_mode = Some(CommandModeContext {
-            selected_text,
-            llm_config,
-        });
-    }
-
-    if context.notify_state() {
-        send_notification(
-            "whisrs",
-            "Command mode: speak your instruction... (press again to stop)",
-        );
-    }
-
-    // Spawn background task: collect audio (with auto-stop), then process.
-    let ds_ref = Arc::clone(&daemon_state);
-    let ctx = Arc::clone(&context);
-    tokio::spawn(async move {
-        command_mode_background(audio_rx, ds_ref, ctx).await;
-    });
-
-    Response::Ok {
-        state: State::Recording,
-    }
-}
-
-/// Background task: collects audio until channel closes (manual stop or auto-stop),
-/// then transcribes the instruction, sends to LLM, and injects the result.
-///
-/// Thin wrapper: collects the recording and takes the session context, runs
-/// the fallible pipeline in [`command_mode_background_inner`], then funnels
-/// every outcome through one finalize block.
-async fn command_mode_background(
-    audio_rx: Option<tokio::sync::mpsc::UnboundedReceiver<Vec<i16>>>,
-    daemon_state: Arc<Mutex<DaemonState>>,
-    context: Arc<DaemonContext>,
-) {
-    let silence_timeout = context.config.general.silence_timeout_ms;
-    let mut auto_stop = AutoStopDetector::new(SILENCE_RMS_THRESHOLD, silence_timeout, SAMPLE_RATE);
-    let mut all_samples: Vec<i16> = Vec::new();
-
-    // Collect audio until silence auto-stop or channel close (manual stop).
-    if let Some(mut rx) = audio_rx {
-        while let Some(chunk) = rx.recv().await {
-            all_samples.extend_from_slice(&chunk);
-            if auto_stop.feed(&chunk) {
-                info!("command mode: silence auto-stop");
-                // Stop capture.
-                let mut ds = daemon_state.lock().await;
-                if let Some(mut capture) = ds.audio_capture.take() {
-                    capture.stop();
-                    tokio::task::spawn_blocking(move || drop(capture));
-                }
-                break;
-            }
-        }
-    }
-
-    // Take the command mode context and transition to transcribing. The lock
-    // is released before the slow transcribe/LLM awaits in the inner fn.
-    let cmd_ctx = {
-        let mut ds = daemon_state.lock().await;
-        ds.audio_capture.take(); // ensure capture is dropped
-        let _ = ds.state_machine.transition(Action::Toggle);
-        ds.command_mode.take()
-    };
-
-    if let Some(cmd_ctx) = cmd_ctx {
-        if let Err(e) = command_mode_background_inner(&all_samples, cmd_ctx, &context).await {
-            let friendly = format_api_error(&e);
-            error!("command mode: {e:#}");
-            if context.notify_error() {
-                send_notification("whisrs", &format!("Command failed: {friendly}"));
-            }
-        }
-    } else {
-        warn!("command mode: context missing, aborting");
-    }
-
-    // Single finalize path — success, benign skip, and error all land here.
-    // Deliberate improvement over the old per-site unwinds: the state
-    // broadcast + overlay reset now run on error paths too, so the tray can
-    // no longer be left showing a stale state after a failure.
-    let mut ds = daemon_state.lock().await;
-    let _ = ds.state_machine.transition(Action::TranscriptionDone);
-    let _ = context.state_tx.send(ds.state_machine.state());
-    if let Some(level_tx) = &context.overlay_level_tx {
-        let _ = level_tx.send(0.0);
-    }
-}
-
-/// Fallible body of [`command_mode_background`]: stop-feedback, gate +
-/// transcribe (via [`transcribe_batch_audio`]), LLM rewrite, inject.
-///
-/// Early `Ok(())` returns are benign skips whose toast the helper already
-/// sent; `Err` is a real failure the outer wrapper formats
-/// (`format_api_error`) and reports. Takes no daemon-state lock, so nothing
-/// is held across the transcribe/LLM awaits.
-async fn command_mode_background_inner(
-    all_samples: &[i16],
-    cmd_ctx: CommandModeContext,
-    context: &DaemonContext,
-) -> Result<()> {
-    if context.config.general.audio_feedback {
-        feedback::play_stop(context.config.general.audio_feedback_volume);
-    }
-
-    let instruction = transcribe_batch_audio(
-        all_samples,
-        context,
-        &context.config.general.language,
-        &BatchOptions {
-            use_prompt: false,
-            save_recovery: false,
-            check_prompt_echo: false,
-            apply_filler: false,
-            origin: BatchOrigin::CommandMode,
-        },
-    )
-    .await?;
-
-    if instruction.is_empty() {
-        // Gated or unintelligible — the helper already toasted.
-        return Ok(());
-    }
-
-    info!("command mode: instruction = {:?}", instruction);
-
-    // Send to LLM.
-    let result =
-        llm::rewrite_text(&cmd_ctx.llm_config, &cmd_ctx.selected_text, &instruction).await?;
-
-    // Inject the result at the cursor through the same policy wrapper the
-    // dictation path uses. By default that means typing keystrokes through
-    // the evdev / Wayland virtual-keyboard pipeline: the AltGr / dead-key
-    // work in xkb-type covers accented and non-ASCII output end-to-end,
-    // including in terminals where Ctrl+V is interpreted as a control
-    // character rather than paste, so the result never touches the clipboard.
-    // (The capture side may still fall back to a simulated Ctrl+C when the
-    // primary selection is empty; see `capture_selection`.)
-    //
-    // `[input] paste = true` switches command mode to clipboard paste for the
-    // same reason it does for dictation: on compositors without the Wayland
-    // virtual-keyboard protocol (e.g. KWin) uinput keycodes are decoded
-    // through the target window's active XKB layout and can come out garbled,
-    // and the clipboard is layout-independent. `is_terminal` only picks
-    // Ctrl+Shift+V over Ctrl+V there; nothing else keys off it. It can only
-    // ever be true under Hyprland and Niri: `get_focused_window_class()`
-    // defaults to `None` for every other tracker (KWin, GNOME, Sway, X11),
-    // so terminals get plain Ctrl+V there. Same gap as the dictation path
-    // above; tracked in issues #70 and #71.
-    //
-    // GUI apps: typing (or pasting) while text is selected replaces the
-    // selection. That is text-widget behavior in GTK, Qt and Electron, and it
-    // holds on Wayland as well, so nothing extra is needed there.
-    //
-    // Terminals: a mouse highlight is a visual overlay, not an editable
-    // selection, so injecting at the cursor would *append* to the line that is
-    // already at the prompt. We clear it first with Ctrl+A / Ctrl+K, then
-    // inject, so the result replaces the highlighted command instead of being
-    // tacked onto it. The clear is best-effort: if it fails we still inject,
-    // because appending the LLM result beats losing it.
-    //
-    // `is_terminal` can only ever be true where
-    // `WindowTracker::get_focused_window_class()` is actually implemented,
-    // which is Hyprland (`src/window/hyprland.rs:59`) and Niri
-    // (`src/window/niri.rs:79`). `src/window/mod.rs:23` defaults it to `None`,
-    // so on KWin, GNOME, Sway and X11 it stays false and terminals there fall
-    // back to plain injection at the cursor (and to plain Ctrl+V on the paste
-    // branch). Tracked in issues #70 and #71; not fixed here.
-    info!("command mode: injecting {} chars", result.len());
-    let text_clone = result.clone();
-    let key_delay = std::time::Duration::from_millis(context.config.input.key_delay_ms);
-    let injector_backend = context.config.input.backend;
-    let paste = context.config.input.paste;
-    let is_terminal = context
-        .window_tracker
-        .get_focused_window_class()
-        .map(|c| is_terminal_class(&c))
-        .unwrap_or(false);
-    match tokio::task::spawn_blocking(move || {
-        if is_terminal {
-            if let Err(e) = clear_line_via_keyboard(key_delay, injector_backend) {
-                warn!("command mode: failed to clear terminal line, injecting anyway: {e:#}");
-            }
-        }
-        inject_text(&text_clone, is_terminal, key_delay, injector_backend, paste)
-    })
-    .await
-    {
-        Ok(Ok(())) => {}
-        Ok(Err(e)) => warn!("command mode: failed to inject text: {e:#}"),
-        Err(e) => warn!("command mode: injection task panicked: {e}"),
-    }
-
-    if context.config.general.audio_feedback {
-        feedback::play_done(context.config.general.audio_feedback_volume);
-    }
-    if context.notify_state() {
-        send_notification("whisrs", &format!("Command applied: {instruction}"));
-    }
-
-    Ok(())
-}
-
-/// Handle a named `[[llm_commands]]` hotkey. A toggle-recording flavor of
-/// plain dictation: first press starts recording, second press stops it and
-/// runs the transcribed text through the command's fixed instruction via the
-/// LLM before typing the result at the cursor. Unlike command mode, there's
-/// no selection or clipboard involved.
-async fn handle_llm_command(
-    daemon_state: Arc<Mutex<DaemonState>>,
-    context: Arc<DaemonContext>,
-    name: String,
-) -> Response {
-    let current_state = {
-        let ds = daemon_state.lock().await;
-        ds.state_machine.state()
-    };
-
-    match current_state {
-        State::Recording => {
-            // Second press: only stop if an llm-command session is active —
-            // mirrors handle_command_mode's own-flag check.
-            let is_llm_command = {
-                let ds = daemon_state.lock().await;
-                ds.llm_command.is_some()
-            };
-            if !is_llm_command {
-                return Response::Error {
-                    message: "recording is active but not for an llm-command — use toggle, \
-                              command, or cancel"
-                        .to_string(),
-                };
-            }
-            let mut ds = daemon_state.lock().await;
-            if let Some(mut capture) = ds.audio_capture.take() {
-                capture.stop();
-                tokio::task::spawn_blocking(move || drop(capture));
-            }
-            info!("llm-command '{name}': manual stop");
-            Response::Ok {
-                state: State::Recording,
-            }
-        }
-        State::Idle => {
-            let entry = context
-                .config
-                .llm_commands
-                .iter()
-                .find(|e| e.name == name)
-                .cloned();
-            match entry {
-                Some(entry) => llm_command_start(daemon_state, context, entry).await,
-                None => Response::Error {
-                    message: format!("no llm_commands entry named '{name}' — check config.toml"),
-                },
-            }
-        }
-        State::Transcribing => Response::Error {
-            message: "cannot start llm-command while transcribing".to_string(),
-        },
-        State::Synthesizing | State::Speaking => Response::Error {
-            message: "cannot start llm-command while reading aloud — cancel read-aloud first"
-                .to_string(),
-        },
-    }
-}
-
-/// Reprogram a named LLM command from the current selection (its `set_hotkey`).
-///
-/// Synchronous: no recording, no LLM, no typing. Captures the highlighted text
-/// and stores it as the command's new instruction — applied live via the
-/// daemon override map and persisted to `config.toml` for the next start.
-async fn handle_set_llm_instruction(
-    daemon_state: Arc<Mutex<DaemonState>>,
-    context: Arc<DaemonContext>,
-    name: String,
-) -> Response {
-    // Only when idle — don't interfere with an active recording / read-aloud.
-    let state = {
-        let ds = daemon_state.lock().await;
-        ds.state_machine.state()
-    };
-    if state != State::Idle {
-        return Response::Error {
-            message: format!("cannot set an instruction while {state:?} — finish or cancel first"),
-        };
-    }
-
-    // The command must exist (its hotkey was registered from config).
-    if !context.config.llm_commands.iter().any(|e| e.name == name) {
-        return Response::Error {
-            message: format!("no llm_commands entry named '{name}' — check config.toml"),
-        };
-    }
-
-    let Some(instruction) = acquire_selected_text(&context).await else {
-        // Not gated by `notify_state()`: this path never records, so the
-        // overlay (which would otherwise justify suppressing toasts) never
-        // shows — without a toast there'd be no feedback at all.
-        if context.notify {
-            send_notification(
-                "whisrs",
-                &format!("'{name}': select the instruction text first"),
-            );
-        }
-        return Response::Ok { state: State::Idle };
-    };
-
-    {
-        let mut ds = daemon_state.lock().await;
-        ds.llm_instruction_overrides
-            .insert(name.clone(), instruction.clone());
-    }
-
-    if let Err(e) = persist_llm_instruction(&name, &instruction) {
-        warn!("llm-command '{name}': failed to persist new instruction to config: {e:#}");
-    }
-
-    info!(
-        "llm-command '{name}': instruction reprogrammed ({} chars)",
-        instruction.len()
-    );
-    // Audible + toast confirmation. Not gated by `notify_state()` (see above):
-    // the set path shows no overlay, so this is the only feedback the user gets.
-    if context.config.general.audio_feedback {
-        feedback::play_done(context.config.general.audio_feedback_volume);
-    }
-    if context.notify {
-        let preview = truncate_preview(&instruction, 77);
-        send_notification("whisrs", &format!("'{name}' set to: {preview}"));
-    }
-
-    Response::Ok { state: State::Idle }
-}
-
-/// Capture the currently selected text: primary selection first, else a
-/// simulated copy (Ctrl+C, or Ctrl+Shift+C in terminals). Restores the
-/// clipboard afterwards (this path never pastes). Returns `None` when nothing
-/// usable is selected. Mirrors command mode's acquisition, minus the recording.
-async fn acquire_selected_text(context: &DaemonContext) -> Option<String> {
-    let clipboard = xkb_type::default_clipboard();
-    let saved = clipboard.get_text().unwrap_or_default();
-
-    let is_terminal = context
-        .window_tracker
-        .get_focused_window_class()
-        .map(|c| is_terminal_class(&c))
-        .unwrap_or(false);
-
-    let mut text = clipboard.get_primary_selection().unwrap_or_default();
-
-    if text.trim().is_empty() || text == saved {
-        let copy_fn = if is_terminal {
-            simulate_terminal_copy
-        } else {
-            simulate_copy
-        };
-        let _ = tokio::task::spawn_blocking(copy_fn).await;
-        tokio::time::sleep(std::time::Duration::from_millis(100)).await;
-        text = clipboard.get_text().unwrap_or_default();
-    }
-
-    // Restore the user's clipboard — the set path never pastes.
-    let _ = clipboard.set_text(&saved);
-
-    let text = text.trim().to_string();
-    if text.is_empty() || text == saved.trim() {
-        None
-    } else {
-        Some(text)
-    }
-}
-
-/// Persist a reprogrammed instruction to `config.toml`: re-read the on-disk
-/// config, update the matching entry, write it back. Best-effort — the
-/// in-memory override already applies to the running daemon.
-fn persist_llm_instruction(name: &str, instruction: &str) -> anyhow::Result<()> {
-    let path = whisrs::config_path();
-    let contents = std::fs::read_to_string(&path)
-        .with_context(|| format!("failed to read config at {}", path.display()))?;
-    let mut config: Config =
-        toml::from_str(&contents).context("failed to parse config for instruction update")?;
-    let entry = config
-        .llm_commands
-        .iter_mut()
-        .find(|e| e.name == name)
-        .ok_or_else(|| anyhow::anyhow!("entry '{name}' not present in config file"))?;
-    entry.instruction = instruction.to_string();
-    whisrs::config::setup::write_config(&config)?;
-    Ok(())
-}
-
-/// Named LLM command, first press: capture the focused window, start
-/// recording, spawn the background processor. Mirrors `handle_toggle`'s Idle
-/// branch (batch path only — this feature doesn't support streaming
-/// backends, same as command mode).
-async fn llm_command_start(
-    daemon_state: Arc<Mutex<DaemonState>>,
-    context: Arc<DaemonContext>,
-    entry: llm::LlmCommandConfig,
-) -> Response {
-    let llm_config = context.config.llm.clone().unwrap_or_default();
-
-    // Capture focused window before recording, like plain dictation — the
-    // result is typed at the cursor, not pasted over a selection.
-    let window_id = match context.window_tracker.get_focused_window() {
-        Ok(id) => Some(id),
-        Err(e) => {
-            warn!("failed to capture focused window: {e}");
-            None
-        }
-    };
-
-    if context.config.general.audio_feedback {
-        feedback::play_start(context.config.general.audio_feedback_volume);
-    }
-
-    let mut capture =
-        match AudioCaptureHandle::start_with_level_tx(context.overlay_level_tx.clone()) {
-            Ok(c) => c,
-            Err(e) => {
-                let msg = format!("{e}");
-                let friendly = if msg.contains("no default audio input device") {
-                    format_no_microphone_error()
-                } else {
-                    format!("Failed to start audio capture: {e}")
-                };
-                error!("{friendly}");
-                return Response::Error { message: friendly };
-            }
-        };
-
-    let audio_rx = capture.take_receiver();
-
-    {
-        let mut ds = daemon_state.lock().await;
-        if let Err(e) = ds.state_machine.transition(Action::Toggle) {
-            return Response::Error {
-                message: format!("state transition failed: {e}"),
-            };
-        }
-        ds.audio_capture = Some(capture);
-        ds.recording_window_id = window_id;
-        ds.recording_started_at = Some(std::time::Instant::now());
-        // Prefer a runtime override set via `set_hotkey`; else the configured
-        // instruction.
-        let instruction = ds
-            .llm_instruction_overrides
-            .get(&entry.name)
-            .cloned()
-            .unwrap_or_else(|| entry.instruction.clone());
-        ds.llm_command = Some(LlmCommandContext {
-            name: entry.name.clone(),
-            instruction,
-            llm_config,
-        });
-    }
-
-    if context.notify_state() {
-        send_notification(
-            "whisrs",
-            &format!("Recording for '{}'... (press again to stop)", entry.name),
-        );
-    }
-
-    let ds_ref = Arc::clone(&daemon_state);
-    let ctx = Arc::clone(&context);
-    tokio::spawn(async move {
-        llm_command_background(audio_rx, ds_ref, ctx).await;
-    });
-
-    Response::Ok {
-        state: State::Recording,
-    }
-}
-
-/// Background task: collects audio until channel closes (manual stop or
-/// auto-stop), transcribes it, applies the command's fixed instruction via
-/// the LLM (reusing the same `llm::rewrite_text` call as command mode, just
-/// with the roles swapped — the dictated text is the "selected text" and the
-/// preset instruction is the "voice instruction"), and types the result at
-/// the cursor.
-///
-/// Thin wrapper: collects the recording and takes the session context, runs
-/// the fallible pipeline in [`llm_command_background_inner`], then funnels
-/// every outcome through one finalize block.
-async fn llm_command_background(
-    audio_rx: Option<tokio::sync::mpsc::UnboundedReceiver<Vec<i16>>>,
-    daemon_state: Arc<Mutex<DaemonState>>,
-    context: Arc<DaemonContext>,
-) {
-    let silence_timeout = context.config.general.silence_timeout_ms;
-    let mut auto_stop = AutoStopDetector::new(SILENCE_RMS_THRESHOLD, silence_timeout, SAMPLE_RATE);
-    let mut all_samples: Vec<i16> = Vec::new();
-
-    if let Some(mut rx) = audio_rx {
-        while let Some(chunk) = rx.recv().await {
-            all_samples.extend_from_slice(&chunk);
-            if auto_stop.feed(&chunk) {
-                info!("llm-command: silence auto-stop");
-                let mut ds = daemon_state.lock().await;
-                if let Some(mut capture) = ds.audio_capture.take() {
-                    capture.stop();
-                    tokio::task::spawn_blocking(move || drop(capture));
-                }
-                break;
-            }
-        }
-    }
-
-    // Take the session context and transition to transcribing. The lock is
-    // released before the slow transcribe/LLM awaits in the inner fn.
-    let (cmd_ctx, window_id, recording_started_at) = {
-        let mut ds = daemon_state.lock().await;
-        ds.audio_capture.take();
-        let _ = ds.state_machine.transition(Action::Toggle);
-        (
-            ds.llm_command.take(),
-            ds.recording_window_id.take(),
-            ds.recording_started_at.take(),
-        )
-    };
-
-    if let Some(cmd_ctx) = cmd_ctx {
-        let name = cmd_ctx.name.clone();
-        if let Err(e) = llm_command_background_inner(
-            &all_samples,
-            cmd_ctx,
-            window_id,
-            recording_started_at,
-            &context,
-        )
-        .await
-        {
-            let friendly = format_api_error(&e);
-            error!("llm-command '{name}': {e:#}");
-            if context.notify_error() {
-                send_notification("whisrs", &format!("'{name}' failed: {friendly}"));
-            }
-        }
-    } else {
-        warn!("llm-command: context missing, aborting");
-    }
-
-    // Single finalize path — success, benign skip, and error all land here.
-    // As in `command_mode_background`, the state broadcast + overlay reset
-    // now deliberately run on error paths too (they used to be success-only),
-    // so the tray can no longer be left showing a stale state after a failure.
-    let mut ds = daemon_state.lock().await;
-    let _ = ds.state_machine.transition(Action::TranscriptionDone);
-    let _ = context.state_tx.send(ds.state_machine.state());
-    if let Some(level_tx) = &context.overlay_level_tx {
-        let _ = level_tx.send(0.0);
-    }
-}
-
-/// Fallible body of [`llm_command_background`]: stop-feedback, gate +
-/// transcribe (via [`transcribe_batch_audio`]), LLM rewrite, inject, history.
-///
-/// Early `Ok(())` returns are benign skips that already surfaced their own
-/// toast; `Err` is a real failure the outer wrapper formats
-/// (`format_api_error`) and reports. Takes no daemon-state lock, so nothing
-/// is held across the transcribe/LLM awaits.
-async fn llm_command_background_inner(
-    all_samples: &[i16],
-    cmd_ctx: LlmCommandContext,
-    window_id: Option<String>,
-    recording_started_at: Option<std::time::Instant>,
-    context: &DaemonContext,
-) -> Result<()> {
-    if context.config.general.audio_feedback {
-        feedback::play_stop(context.config.general.audio_feedback_volume);
-    }
-
-    let text = transcribe_batch_audio(
-        all_samples,
-        context,
-        &context.config.general.language,
-        &BatchOptions {
-            use_prompt: true,
-            save_recovery: false,
-            // The transcript is LLM input here: an echoed prompt would be
-            // silently rewritten into plausible garbage, so the echo check
-            // matters just as much as on the dictation path.
-            check_prompt_echo: true,
-            apply_filler: false,
-            origin: BatchOrigin::LlmCommand {
-                name: &cmd_ctx.name,
-            },
-        },
-    )
-    .await?;
-
-    if text.is_empty() {
-        // Gated, echo-dropped, or unintelligible — the helper already toasted.
-        return Ok(());
-    }
-
-    info!(
-        "llm-command '{}': transcribed {} chars",
-        cmd_ctx.name,
-        text.len()
-    );
-
-    let result = llm::rewrite_text(&cmd_ctx.llm_config, &text, &cmd_ctx.instruction).await?;
-
-    if result.is_empty() {
-        if context.notify_error() {
-            send_notification(
-                "whisrs",
-                &format!("'{}': LLM returned empty text", cmd_ctx.name),
-            );
-        }
-        return Ok(());
-    }
-
-    // Restore window focus, then inject the result at the cursor — keystrokes,
-    // or clipboard paste when `[input] paste = true` (layout-independent).
-    if let Some(wid) = &window_id {
-        if let Err(e) = context.window_tracker.focus_window(wid) {
-            warn!("failed to restore window focus: {e}");
-        } else {
-            tokio::time::sleep(std::time::Duration::from_millis(100)).await;
-        }
-    }
-
-    let result_clone = result.clone();
-    let key_delay = std::time::Duration::from_millis(context.config.input.key_delay_ms);
-    let injector_backend = context.config.input.backend;
-    let paste = context.config.input.paste;
-    let is_terminal = if paste {
-        context
-            .window_tracker
-            .get_focused_window_class()
-            .map(|c| is_terminal_class(&c))
-            .unwrap_or(false)
-    } else {
-        false
-    };
-    match tokio::task::spawn_blocking(move || {
-        inject_text(
-            &result_clone,
-            is_terminal,
-            key_delay,
-            injector_backend,
-            paste,
-        )
-    })
-    .await
-    {
-        Ok(Ok(())) => {}
-        Ok(Err(e)) => warn!(
-            "llm-command '{}': failed to inject text: {e:#}",
-            cmd_ctx.name
-        ),
-        Err(e) => warn!(
-            "llm-command '{}': failed to join injection task: {e}",
-            cmd_ctx.name
-        ),
-    }
-
-    let duration_secs = recording_started_at
-        .map(|t| t.elapsed().as_secs_f64())
-        .unwrap_or(0.0);
-    save_history_entry(
-        &result,
-        &format!("llm:{}", cmd_ctx.name),
-        &context.config.general.language,
-        duration_secs,
-    );
-
-    if context.config.general.audio_feedback {
-        feedback::play_done(context.config.general.audio_feedback_volume);
-    }
-    if context.notify_state() {
-        let preview = truncate_preview(&result, 77);
-        send_notification("whisrs", &format!("'{}' done: {preview}", cmd_ctx.name));
-    }
-
-    Ok(())
-}
-
-/// Simulate a key combo (e.g. Ctrl+C, Ctrl+V) via a temporary uinput device.
-fn simulate_key_combo(modifier: evdev::Key, key: evdev::Key) -> anyhow::Result<()> {
-    use evdev::{AttributeSet, EventType, InputEvent, Key};
-    use std::thread;
-    use std::time::Duration;
-
-    let mut keys = AttributeSet::<Key>::new();
-    keys.insert(modifier);
-    keys.insert(key);
-
-    let mut device = evdev::uinput::VirtualDeviceBuilder::new()
-        .context("failed to create VirtualDeviceBuilder")?
-        .name("whisrs command")
-        .with_keys(&keys)
-        .context("failed to register key events")?
-        .build()
-        .context("failed to build uinput device")?;
-
-    thread::sleep(Duration::from_millis(200));
-
-    // Press modifier.
-    device.emit(&[InputEvent::new(EventType::KEY, modifier.code(), 1)])?;
-    thread::sleep(Duration::from_millis(2));
-    // Press key.
-    device.emit(&[InputEvent::new(EventType::KEY, key.code(), 1)])?;
-    thread::sleep(Duration::from_millis(2));
-    // Release key.
-    device.emit(&[InputEvent::new(EventType::KEY, key.code(), 0)])?;
-    thread::sleep(Duration::from_millis(2));
-    // Release modifier.
-    device.emit(&[InputEvent::new(EventType::KEY, modifier.code(), 0)])?;
-    thread::sleep(Duration::from_millis(2));
-
-    Ok(())
-}
-
-/// Simulate a two-modifier + key combo (e.g. Ctrl+Shift+V) via a temporary uinput device.
-fn simulate_key_combo_2mod(
-    mod1: evdev::Key,
-    mod2: evdev::Key,
-    key: evdev::Key,
-) -> anyhow::Result<()> {
-    use evdev::{AttributeSet, EventType, InputEvent, Key};
-    use std::thread;
-    use std::time::Duration;
-
-    let mut keys = AttributeSet::<Key>::new();
-    keys.insert(mod1);
-    keys.insert(mod2);
-    keys.insert(key);
-
-    let mut device = evdev::uinput::VirtualDeviceBuilder::new()
-        .context("failed to create VirtualDeviceBuilder")?
-        .name("whisrs command")
-        .with_keys(&keys)
-        .context("failed to register key events")?
-        .build()
-        .context("failed to build uinput device")?;
-
-    thread::sleep(Duration::from_millis(200));
-
-    device.emit(&[InputEvent::new(EventType::KEY, mod1.code(), 1)])?;
-    thread::sleep(Duration::from_millis(2));
-    device.emit(&[InputEvent::new(EventType::KEY, mod2.code(), 1)])?;
-    thread::sleep(Duration::from_millis(2));
-    device.emit(&[InputEvent::new(EventType::KEY, key.code(), 1)])?;
-    thread::sleep(Duration::from_millis(2));
-    device.emit(&[InputEvent::new(EventType::KEY, key.code(), 0)])?;
-    thread::sleep(Duration::from_millis(2));
-    device.emit(&[InputEvent::new(EventType::KEY, mod2.code(), 0)])?;
-    thread::sleep(Duration::from_millis(2));
-    device.emit(&[InputEvent::new(EventType::KEY, mod1.code(), 0)])?;
-    thread::sleep(Duration::from_millis(2));
-
-    Ok(())
-}
-
-/// Simulate Ctrl+C (copy) via uinput.
-fn simulate_copy() -> anyhow::Result<()> {
-    simulate_key_combo(evdev::Key::KEY_LEFTCTRL, evdev::Key::KEY_C)
-}
-
-/// Simulate Ctrl+Shift+C (terminal copy) via uinput.
-fn simulate_terminal_copy() -> anyhow::Result<()> {
-    simulate_key_combo_2mod(
-        evdev::Key::KEY_LEFTCTRL,
-        evdev::Key::KEY_LEFTSHIFT,
-        evdev::Key::KEY_C,
-    )
-}
-
-/// Known terminal emulator window classes (lowercase for matching).
-const TERMINAL_CLASSES: &[&str] = &[
-    "alacritty",
-    "kitty",
-    "foot",
-    "wezterm",
-    "gnome-terminal",
-    "konsole",
-    "xterm",
-    "urxvt",
-    "terminator",
-    "tilix",
-    "st",
-    "xfce4-terminal",
-    "sakura",
-    "guake",
-    "yakuake",
-    "termite",
-    "cool-retro-term",
-    "ghostty",
-];
-
-/// Check if a window class corresponds to a terminal emulator.
-fn is_terminal_class(class: &str) -> bool {
-    let lower = class.to_lowercase();
-    TERMINAL_CLASSES.iter().any(|t| lower.contains(t))
-}
-
-/// Returns `true` when the first character of `text` is a punctuation mark
-/// that should not have a space inserted before it.
-///
-/// Covers ASCII sentence/clause terminators and closing brackets, their CJK
-/// fullwidth equivalents, the ellipsis character, and closing curly quotes.
-/// Opening brackets, opening curly quotes, and dashes are intentionally
-/// excluded — they should retain a leading space.
-fn leads_with_punct(text: &str) -> bool {
-    text.starts_with([
-        // ASCII
-        '.', ',', '!', '?', ';', ':', ')', ']', '}', // CJK fullwidth equivalents
-        '。', '，', '！', '？', '；', '：', '）', '】', '｝', // Ellipsis
-        '…',  // Closing curly quotes (U+2019, U+201D)
-        '\u{2019}', '\u{201D}',
-    ])
-}
-
-/// Capture the currently-selected text for read-aloud and command mode.
-///
-/// The primary selection (highlighted text) is authoritative when present: it
-/// needs no key simulation, so a non-empty primary selection is returned
-/// directly. This also avoids the clipboard-equality heuristic below — which
-/// only makes sense in the copy-fallback path — wrongly rejecting a selection
-/// that happens to equal the current clipboard.
-///
-/// When the primary selection is empty, fall back to a simulated Ctrl+C
-/// (Ctrl+Shift+C in terminals) and read the clipboard. A short settle delay
-/// precedes the simulated copy: the triggering hotkey is typically a modifier
-/// combo (e.g. Alt+Shift+A), and if the user is still physically holding those
-/// modifiers when Ctrl+C fires, the app receives a garbled combo and the copy
-/// silently fails. The delay lets a briefly-held hotkey clear first.
-///
-/// That fallback overwrites the user's clipboard, so it is restored as soon as
-/// the copied text has been read — synchronously, because nothing downstream
-/// needs the clipboard to keep holding the capture (command mode injects the
-/// LLM result, read-aloud speaks the text). The restore is skipped, rather
-/// than clobbering the clipboard, under the same two conditions as
-/// [`inject_text`]: the original clipboard wasn't readable as text (an image
-/// or a file list), or something else copied over our simulated Ctrl+C in the
-/// meantime. It is also skipped when the copy captured nothing and left the
-/// clipboard unchanged, so nothing is written when there was no selection to
-/// copy. A failed restore is logged, never fatal.
-///
-/// Returns `Ok(text)` on success, or `Err(message)` describing why nothing was
-/// captured (caller surfaces this as a `Response::Error` + notification).
-async fn capture_selection(context: &DaemonContext) -> Result<String, String> {
-    let clipboard = xkb_type::default_clipboard();
-
-    // Trust a non-empty primary selection directly — no copy, no equality
-    // check, and no clipboard write of any kind.
-    let primary = clipboard.get_primary_selection().unwrap_or_default();
-    if !primary.is_empty() {
-        return Ok(primary);
-    }
-
-    // No primary selection: fall back to a simulated copy + clipboard read.
-    // `None` means the clipboard held something that isn't text.
-    let saved_clipboard = clipboard.get_text().ok();
-
-    let is_terminal = context
-        .window_tracker
-        .get_focused_window_class()
-        .map(|c| is_terminal_class(&c))
-        .unwrap_or(false);
-
-    // Let any briefly-held hotkey modifiers clear before synthesizing Ctrl+C.
-    tokio::time::sleep(std::time::Duration::from_millis(250)).await;
-
-    let copy_fn = if is_terminal {
-        simulate_terminal_copy
-    } else {
-        simulate_copy
-    };
-
-    match tokio::task::spawn_blocking(copy_fn).await {
-        Ok(Ok(())) => {}
-        Ok(Err(e)) => return Err(format!("failed to copy selection: {e}")),
-        Err(e) => return Err(format!("copy task panicked: {e}")),
-    }
-
-    tokio::time::sleep(std::time::Duration::from_millis(100)).await;
-    let copied = clipboard
-        .get_text()
-        .map_err(|e| format!("failed to read clipboard: {e}"))?;
-
-    // The copy clobbered the user's clipboard with the selection; put the
-    // original back now that we've read it (see doc comment above). Skipped
-    // when the original wasn't text, when the clipboard no longer holds what
-    // our Ctrl+C put there, and when the copy left the clipboard exactly as it
-    // was (nothing was selected, so there is nothing to undo). Writing an
-    // identical value back is not free: on Wayland it takes selection
-    // ownership and re-offers plain text only, downgrading a clipboard that
-    // held text/html and giving an empty one an empty-string owner.
-    let saved_to_restore = saved_clipboard
-        .as_deref()
-        .filter(|saved| *saved != copied.as_str());
-    if let Some(saved) = saved_to_restore {
-        match clipboard.get_text() {
-            Ok(current) if current == copied => {
-                if let Err(e) = clipboard.set_text(saved) {
-                    warn!("failed to restore clipboard after selection capture: {e}");
-                }
-            }
-            Ok(_) => {
-                debug!("clipboard changed during selection capture; skipping restore");
-            }
-            Err(e) => {
-                warn!("failed to read clipboard before restore, skipping restore: {e}");
-            }
-        }
-    }
-
-    // With no primary selection, an unchanged clipboard means the copy captured
-    // nothing (nothing selected, or the held hotkey garbled Ctrl+C).
-    if copied.is_empty() || saved_clipboard.as_deref() == Some(copied.as_str()) {
-        return Err("no text selected — select some text first".to_string());
-    }
-
-    Ok(copied)
-}
-
-/// Read the selected text aloud via TTS.
-///
-/// FSM-authoritative: read-aloud has its own `Synthesizing`/`Speaking` states,
-/// so "is read-aloud active" is determined by [`State`], not the `tts_stop`
-/// flag (which now exists purely as the low-level playback interrupt). This
-/// removes the race where a `Speak` press landing in the gap between playback
-/// finishing and the cleanup task clearing `tts_stop` was swallowed as a stop.
-///
-/// The daemon mutex is never held across the synth or playback `.await`.
-async fn handle_speak(
-    daemon_state: Arc<Mutex<DaemonState>>,
-    context: Arc<DaemonContext>,
-) -> Response {
-    use std::sync::atomic::{AtomicBool, Ordering};
-
-    // (a) Inspect the FSM. A second press while read-aloud is active is a
-    // deterministic stop, regardless of playback timing.
-    {
-        let mut ds = daemon_state.lock().await;
-        match ds.state_machine.state() {
-            State::Speaking | State::Synthesizing => {
-                if let Some(stop) = ds.tts_stop.take() {
-                    stop.store(true, Ordering::Release);
-                }
-                let _ = ds.state_machine.transition(Action::Cancel);
-                let new_state = ds.state_machine.state();
-                let _ = context.state_tx.send(new_state);
-                if let Some(level_tx) = &context.overlay_level_tx {
-                    let _ = level_tx.send(0.0);
-                }
-                info!("speak: stopped in-progress read-aloud");
-                return Response::Ok { state: new_state };
-            }
-            State::Recording | State::Transcribing => {
-                return Response::Error {
-                    message: "busy — finish or cancel recording first".to_string(),
-                };
-            }
-            State::Idle => {
-                // Fall through; do NOT transition yet.
-            }
-        }
-    }
-
-    // (b) Verify TTS is configured/enabled and build the backend. State is
-    // still Idle on any error here — nothing to unwind.
-    let tts_config = match &context.config.tts {
-        Some(t) if t.enabled => t.clone(),
-        Some(_) => {
-            return Response::Error {
-                message: "TTS is disabled — set [tts] enabled = true in config.toml".to_string(),
-            };
-        }
-        None => {
-            return Response::Error {
-                message: "TTS is not configured — add a [tts] section to config.toml".to_string(),
-            };
-        }
-    };
-
-    let api_key = resolve_tts_api_key(&context.config);
-    let backend = match whisrs::tts::create_backend(&tts_config, api_key) {
-        Ok(b) => b,
-        Err(e) => {
-            return Response::Error {
-                message: e.to_string(),
-            };
-        }
-    };
-
-    // (c) Capture the selection. On failure (incl. empty), surface an error
-    // and — since this is hotkey-triggered — notify so the user sees why.
-    info!("speak: getting selected text");
-    let selected_text = match capture_selection(&context).await {
-        Ok(text) => text,
-        Err(message) => {
-            if context.notify_error() {
-                send_notification("whisrs", &format!("Read-aloud: {message}"));
-            }
-            return Response::Error { message };
-        }
-    };
-
-    // (d) Re-lock; only begin synthesizing if still Idle (a concurrent command
-    // may have intervened). Do NOT proceed otherwise.
-    {
-        let mut ds = daemon_state.lock().await;
-        if ds.state_machine.state() != State::Idle {
-            return Response::Ok {
-                state: ds.state_machine.state(),
-            };
-        }
-        let _ = ds.state_machine.transition(Action::SpeakStart);
-        let _ = context.state_tx.send(ds.state_machine.state());
-    }
-
-    info!("speak: synthesizing {} chars", selected_text.len());
-    if context.notify_state() {
-        send_notification("whisrs", "Reading selection aloud...");
-    }
-
-    // (e) Synthesize (no lock held).
-    let synth_result = backend.synthesize(&selected_text).await;
-
-    // (f) Re-lock to decide whether to play.
-    let final_state = {
-        let mut ds = daemon_state.lock().await;
-
-        // A second press during synthesis cancelled us — state is no longer
-        // Synthesizing. Don't play; report the current state.
-        if ds.state_machine.state() != State::Synthesizing {
-            info!("speak: synthesis superseded by a concurrent command");
-            return Response::Ok {
-                state: ds.state_machine.state(),
-            };
-        }
-
-        let wav_bytes = match synth_result {
-            Ok(bytes) => bytes,
-            Err(e) => {
-                error!("speak: synthesis failed: {e}");
-                let _ = ds.state_machine.transition(Action::SpeakDone);
-                let _ = context.state_tx.send(ds.state_machine.state());
-                if let Some(level_tx) = &context.overlay_level_tx {
-                    let _ = level_tx.send(0.0);
-                }
-                drop(ds);
-                if context.notify_error() {
-                    send_notification("whisrs", &format!("Read-aloud failed: {e}"));
-                }
-                return Response::Error {
-                    message: format!("TTS synthesis failed: {e}"),
-                };
-            }
-        };
-
-        let stop = Arc::new(AtomicBool::new(false));
-        ds.tts_stop = Some(Arc::clone(&stop));
-        let _ = ds.state_machine.transition(Action::SpeakPlaying);
-        let new_state = ds.state_machine.state();
-        let _ = context.state_tx.send(new_state);
-
-        // (g) Spawn interruptible playback feeding the speaking overlay.
-        let ds_ref = Arc::clone(&daemon_state);
-        let context_for_cleanup = Arc::clone(&context);
-        let playback_stop = Arc::clone(&stop);
-        let level_tx = context.overlay_level_tx.clone();
-        tokio::spawn(async move {
-            let result = tokio::task::spawn_blocking(move || {
-                whisrs::audio::playback::play_wav(&wav_bytes, stop, level_tx)
-            })
-            .await;
-            match result {
-                Ok(Ok(())) => debug!("speak: playback finished"),
-                Ok(Err(e)) => warn!("speak: playback failed: {e}"),
-                Err(e) => warn!("speak: playback task panicked: {e}"),
-            }
-
-            // On finish/interrupt: only finalize if we still own playback and
-            // the FSM is still Speaking. A second Speak / Cancel already
-            // transitioned to Idle and replaced/cleared tts_stop.
-            let mut ds = ds_ref.lock().await;
-            let still_ours = ds
-                .tts_stop
-                .as_ref()
-                .is_some_and(|current| Arc::ptr_eq(current, &playback_stop));
-            if ds.state_machine.state() == State::Speaking && still_ours {
-                let _ = ds.state_machine.transition(Action::SpeakDone);
-                ds.tts_stop = None;
-                let _ = context_for_cleanup.state_tx.send(ds.state_machine.state());
-                if let Some(level_tx) = &context_for_cleanup.overlay_level_tx {
-                    let _ = level_tx.send(0.0);
-                }
-            }
-        });
-
-        new_state
-    };
-
-    Response::Ok { state: final_state }
-}
-
-async fn handle_cancel(
-    daemon_state: Arc<Mutex<DaemonState>>,
-    context: Arc<DaemonContext>,
-) -> Response {
-    let mut ds = daemon_state.lock().await;
-
-    // Stop any in-progress TTS playback regardless of recording state.
-    let stopped_tts = if let Some(stop) = ds.tts_stop.take() {
-        stop.store(true, std::sync::atomic::Ordering::Release);
-        info!("cancel: stopped TTS playback");
-        true
-    } else {
-        false
-    };
-
-    match ds.state_machine.transition(Action::Cancel) {
-        Ok(new_state) => {
-            if let Some(mut capture) = ds.audio_capture.take() {
-                capture.stop();
-                tokio::task::spawn_blocking(move || drop(capture));
-            }
-            // Stop the typing loop BEFORE aborting the pipeline. Aborting
-            // only drops the outer pipeline future; the detached backend and
-            // typing tasks keep running, the backend treats the dropped audio
-            // channel as a normal end-of-stream and flushes the trailing
-            // phrase — without this flag the typing task would type it.
-            if let Some(cancel) = ds.streaming_cancel.take() {
-                cancel.store(true, Ordering::SeqCst);
-            }
-            if let Some(task) = ds.streaming_task.take() {
-                task.abort();
-            }
-            ds.recording_window_id = None;
-            ds.session_language = None;
-            if let Some(level_tx) = &context.overlay_level_tx {
-                let _ = level_tx.send(0.0);
-            }
-            info!("cancelled recording");
-            if context.notify_state() {
-                send_notification("whisrs", "Recording cancelled");
-            }
-            Response::Ok { state: new_state }
-        }
-        // Nothing was recording. If we still interrupted TTS playback, that's a
-        // successful cancel — report the current (Idle) state rather than the
-        // invalid-transition error.
-        Err(_) if stopped_tts => Response::Ok {
-            state: ds.state_machine.state(),
-        },
-        Err(e) => Response::Error {
-            message: e.to_string(),
-        },
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -3575,446 +282,5 @@ mod tests {
     #[test]
     fn daemon_parses_with_no_arguments() {
         assert!(Args::try_parse_from(["whisrsd"]).is_ok());
-    }
-
-    #[test]
-    fn resolve_language_prefers_override() {
-        assert_eq!(resolve_language(Some("pl".to_string()), "en"), "pl");
-    }
-
-    #[test]
-    fn resolve_language_falls_back_to_default() {
-        assert_eq!(resolve_language(None, "en"), "en");
-    }
-
-    /// Mirrors handle_toggle's start-press: validation runs first and its
-    /// error response is returned before the Idle→Recording transition, so
-    /// a bad `-l` override never starts a recording.
-    #[test]
-    fn start_press_rejects_invalid_language_override_before_recording() {
-        let ds = DaemonState::new();
-
-        let response = validate_toggle_language(Some("english".to_string())).unwrap_err();
-        assert!(matches!(response, Response::Error { .. }));
-
-        // handle_toggle returns that response without touching the state
-        // machine — the daemon stays idle and no capture is started.
-        assert_eq!(ds.state_machine.state(), State::Idle);
-    }
-
-    /// Valid overrides are normalized (whisper.cpp needs lowercase codes);
-    /// a plain toggle passes through untouched.
-    #[test]
-    fn start_press_normalizes_valid_language_override() {
-        assert_eq!(
-            validate_toggle_language(Some("PL".to_string())).unwrap(),
-            Some("pl".to_string())
-        );
-        assert_eq!(validate_toggle_language(None).unwrap(), None);
-    }
-
-    /// Minimal config with a batch (non-streaming) backend and an `en` default
-    /// language, mirroring the setup where issue reviews caught the batch path
-    /// ignoring `-l`.
-    fn batch_config() -> Config {
-        toml::from_str(
-            r#"
-            [general]
-            backend = "groq"
-            language = "en"
-            "#,
-        )
-        .unwrap()
-    }
-
-    /// The batch path must transcribe with the language the session was
-    /// started with (`toggle -l pl`), not the config default. Mirrors the
-    /// handle_toggle flow: the start-press persists the resolved language in
-    /// `DaemonState`; the stop-press consumes it and builds the batch
-    /// `TranscriptionConfig` from it.
-    #[test]
-    fn batch_stop_uses_session_language_override() {
-        let config = batch_config();
-        let mut ds = DaemonState::new();
-
-        // Start press with `-l pl`: resolve and persist the session language.
-        ds.state_machine.transition(Action::Toggle).unwrap();
-        ds.session_language = Some(resolve_language(
-            Some("pl".to_string()),
-            &config.general.language,
-        ));
-
-        // Stop press: the batch path consumes the stored session language.
-        ds.state_machine.transition(Action::Toggle).unwrap();
-        let session_language = ds.take_session_language(&config.general.language);
-        assert_eq!(session_language, "pl");
-
-        // This is the config `process_recording_batch` sends to the backend
-        // (and the language `save_history_entry` records).
-        let tc = build_transcription_config(&config, &session_language);
-        assert_eq!(tc.language, "pl");
-    }
-
-    /// Consuming the session language ends the language session: the next
-    /// recording must fall back to the config default, not inherit `pl`.
-    #[test]
-    fn session_language_does_not_leak_into_next_session() {
-        let config = batch_config();
-        let mut ds = DaemonState::new();
-
-        ds.session_language = Some("pl".to_string());
-        assert_eq!(ds.take_session_language(&config.general.language), "pl");
-
-        // Session over — a plain toggle now uses the default again.
-        assert_eq!(ds.take_session_language(&config.general.language), "en");
-    }
-
-    /// Without an override the batch path keeps using the config default.
-    #[test]
-    fn batch_stop_defaults_to_config_language() {
-        let config = batch_config();
-        let mut ds = DaemonState::new();
-
-        ds.session_language = Some(resolve_language(None, &config.general.language));
-        let session_language = ds.take_session_language(&config.general.language);
-        let tc = build_transcription_config(&config, &session_language);
-        assert_eq!(tc.language, "en");
-    }
-
-    #[test]
-    fn truncate_ascii_short() {
-        assert_eq!(truncate_preview("hello", 77), "hello");
-    }
-
-    #[test]
-    fn truncate_ascii_exact() {
-        let text = "a".repeat(77);
-        assert_eq!(truncate_preview(&text, 77), text);
-    }
-
-    #[test]
-    fn truncate_ascii_long() {
-        let text = "a".repeat(100);
-        let expected = format!("{}...", "a".repeat(77));
-        assert_eq!(truncate_preview(&text, 77), expected);
-    }
-
-    #[test]
-    fn truncate_cyrillic() {
-        // Cyrillic chars are 2 bytes each. 40 Cyrillic chars = 80 bytes > 77.
-        let text = "а".repeat(40); // 80 bytes
-        let preview = truncate_preview(&text, 77);
-        assert!(preview.ends_with("..."));
-        // Should truncate to 38 chars (76 bytes) + "..." since 39 chars = 78 bytes > 77.
-        assert_eq!(preview, format!("{}...", "а".repeat(38)));
-    }
-
-    #[test]
-    fn truncate_cyrillic_issue_repro() {
-        // Exact text from issue #4.
-        let text = "Вот это рутина, блять, вы дешевые щели, блять.";
-        let preview = truncate_preview(text, 77);
-        // Must not panic and must be valid UTF-8.
-        assert!(preview.is_char_boundary(0));
-        assert!(!preview.is_empty());
-    }
-
-    #[test]
-    fn truncate_cjk() {
-        // CJK chars are 3 bytes each. 27 CJK chars = 81 bytes > 77.
-        let text = "字".repeat(27);
-        let preview = truncate_preview(&text, 77);
-        assert!(preview.ends_with("..."));
-        // 25 chars = 75 bytes (fits), 26 chars = 78 bytes (too big).
-        assert_eq!(preview, format!("{}...", "字".repeat(25)));
-    }
-
-    #[test]
-    fn truncate_arabic() {
-        // Arabic chars are 2 bytes each. 40 Arabic chars = 80 bytes > 77.
-        let text = "ش".repeat(40);
-        let preview = truncate_preview(&text, 77);
-        assert!(preview.ends_with("..."));
-        assert_eq!(preview, format!("{}...", "ش".repeat(38)));
-    }
-
-    #[test]
-    fn truncate_emoji() {
-        // Emoji are 4 bytes each. 20 emoji = 80 bytes > 77.
-        let text = "😀".repeat(20);
-        let preview = truncate_preview(&text, 77);
-        assert!(preview.ends_with("..."));
-        // 19 emoji = 76 bytes (fits).
-        assert_eq!(preview, format!("{}...", "😀".repeat(19)));
-    }
-
-    #[test]
-    fn truncate_mixed_scripts() {
-        // Mix of ASCII (1 byte), Cyrillic (2 bytes), CJK (3 bytes).
-        let text = "Hello Мир 世界! This is a test of mixed script truncation that is long enough";
-        let preview = truncate_preview(text, 77);
-        assert!(preview.ends_with("..."));
-        // Verify it doesn't panic and produces valid UTF-8.
-        assert!(preview.len() <= 80); // 77 + "..."
-    }
-
-    #[test]
-    fn truncate_empty() {
-        assert_eq!(truncate_preview("", 77), "");
-    }
-
-    #[test]
-    fn transcription_prompt_neither() {
-        assert_eq!(transcription_prompt(None, &[]), None);
-    }
-
-    #[test]
-    fn transcription_prompt_vocab_only() {
-        let vocab = vec!["whisrs".to_string(), "Hyprland".to_string()];
-        assert_eq!(
-            transcription_prompt(None, &vocab),
-            Some("whisrs, Hyprland".to_string())
-        );
-    }
-
-    #[test]
-    fn transcription_prompt_prose_only() {
-        assert_eq!(
-            transcription_prompt(Some("Embedded Linux dictation."), &[]),
-            Some("Embedded Linux dictation.".to_string())
-        );
-    }
-
-    #[test]
-    fn transcription_prompt_both_combined_with_separator() {
-        let vocab = vec!["whisrs".to_string(), "Hyprland".to_string()];
-        assert_eq!(
-            transcription_prompt(Some("Embedded Linux dictation"), &vocab),
-            Some("Embedded Linux dictation. whisrs, Hyprland".to_string())
-        );
-    }
-
-    #[test]
-    fn transcription_prompt_blank_prose_treated_as_absent() {
-        let vocab = vec!["whisrs".to_string()];
-        // Whitespace-only prompt must not bleed an empty "" into the output.
-        assert_eq!(
-            transcription_prompt(Some("   \t\n  "), &vocab),
-            Some("whisrs".to_string())
-        );
-    }
-
-    #[test]
-    fn transcription_prompt_empty_string_with_empty_vocab_is_none() {
-        assert_eq!(transcription_prompt(Some(""), &[]), None);
-    }
-
-    #[test]
-    fn leads_with_punct_ascii_terminators_and_closers() {
-        // Sentence/clause terminators and closing brackets that should not get
-        // a leading space inserted before them.
-        for p in [".", ",", "!", "?", ";", ":", ")", "]", "}"] {
-            assert!(
-                leads_with_punct(p),
-                "expected leads_with_punct({p:?}) to be true",
-            );
-        }
-    }
-
-    #[test]
-    fn leads_with_punct_ascii_followed_by_more_text() {
-        // Helper should fire based on the first char, regardless of what follows.
-        assert!(leads_with_punct(". And then"));
-        assert!(leads_with_punct(", and"));
-        assert!(leads_with_punct("?!"));
-    }
-
-    #[test]
-    fn leads_with_punct_unicode_cjk_fullwidth() {
-        // CJK fullwidth equivalents of ASCII punctuation.
-        for p in ["。", "，", "！", "？", "；", "：", "）", "】", "｝"] {
-            assert!(
-                leads_with_punct(p),
-                "expected leads_with_punct({p:?}) to be true",
-            );
-        }
-    }
-
-    #[test]
-    fn leads_with_punct_unicode_ellipsis() {
-        assert!(leads_with_punct("…"));
-        assert!(leads_with_punct("… and then"));
-    }
-
-    #[test]
-    fn leads_with_punct_unicode_closing_curly_quotes() {
-        // U+2019 RIGHT SINGLE QUOTATION MARK
-        assert!(leads_with_punct("\u{2019}"));
-        // U+201D RIGHT DOUBLE QUOTATION MARK
-        assert!(leads_with_punct("\u{201D}"));
-    }
-
-    #[test]
-    fn leads_with_punct_leading_space_is_false() {
-        // Whitespace prefix means the caller already has spacing — do not
-        // suppress an additional one based on punctuation logic.
-        assert!(!leads_with_punct(" hello"));
-        assert!(!leads_with_punct(" ."));
-    }
-
-    #[test]
-    fn leads_with_punct_non_punct_prefix_is_false() {
-        assert!(!leads_with_punct("hello"));
-        assert!(!leads_with_punct("a."));
-        assert!(!leads_with_punct("1, 2, 3"));
-    }
-
-    #[test]
-    fn leads_with_punct_empty_is_false() {
-        assert!(!leads_with_punct(""));
-    }
-
-    #[test]
-    fn leads_with_punct_opening_brackets_are_false() {
-        // Opening brackets should retain a leading space.
-        assert!(!leads_with_punct("("));
-        assert!(!leads_with_punct("["));
-        assert!(!leads_with_punct("{"));
-    }
-
-    #[test]
-    fn leads_with_punct_opening_curly_quotes_are_false() {
-        // U+2018 LEFT SINGLE QUOTATION MARK
-        assert!(!leads_with_punct("\u{2018}"));
-        // U+201C LEFT DOUBLE QUOTATION MARK
-        assert!(!leads_with_punct("\u{201C}"));
-    }
-
-    #[test]
-    fn sanitize_ws_endpoint_for_log_strips_credentials_and_query() {
-        assert_eq!(
-            sanitize_ws_endpoint_for_log("ws://user:secret@localhost:1234/realtime?token=abc"),
-            "ws://localhost:1234/realtime"
-        );
-    }
-
-    #[test]
-    fn get_model_for_openai_compatible_realtime_backend() {
-        let config = Config {
-            general: whisrs::GeneralConfig {
-                backend: "openai-compatible-realtime".to_string(),
-                ..Default::default()
-            },
-            audio: Default::default(),
-            input: Default::default(),
-            deepgram: None,
-            groq: None,
-            openai: None,
-            local_whisper: None,
-            local_vosk: None,
-            local_parakeet: None,
-            asr_sidecar: None,
-            openai_compatible_realtime: Some(whisrs::OpenAiCompatibleRealtimeConfig {
-                url: "ws://localhost:1234/realtime".to_string(),
-                model: "Whisper-Tiny".to_string(),
-                profile: "lemonade".to_string(),
-                turn_detection: "server-vad".to_string(),
-                api_key: None,
-            }),
-            llm: None,
-            hotkeys: None,
-            llm_commands: Vec::new(),
-            overlay: None,
-            tts: None,
-        };
-
-        assert_eq!(get_model_for_backend(&config), "Whisper-Tiny");
-    }
-
-    #[test]
-    fn format_api_error_preserves_realtime_flush_timeout_context() {
-        let err = anyhow::anyhow!(
-            "timed out waiting for final transcription completion from ws://example after commit"
-        );
-        assert_eq!(
-            format_api_error(&err),
-            "Realtime transcription timed out waiting for the server to finish after stop"
-        );
-    }
-
-    /// Spawn `run_typing_batcher` with a recording sink; returns the text
-    /// channel sender, the cancel flag, the sink log, and the join handle.
-    #[allow(clippy::type_complexity)]
-    fn spawn_test_batcher() -> (
-        tokio::sync::mpsc::Sender<String>,
-        Arc<AtomicBool>,
-        Arc<StdMutex<Vec<String>>>,
-        tokio::task::JoinHandle<String>,
-    ) {
-        let (text_tx, text_rx) = tokio::sync::mpsc::channel::<String>(64);
-        let cancel = Arc::new(AtomicBool::new(false));
-        let typed = Arc::new(StdMutex::new(Vec::<String>::new()));
-
-        let sink_log = Arc::clone(&typed);
-        let batcher = tokio::spawn(run_typing_batcher(
-            text_rx,
-            Arc::clone(&cancel),
-            None,
-            move |text| {
-                let sink_log = Arc::clone(&sink_log);
-                async move {
-                    sink_log.lock().unwrap().push(text);
-                }
-            },
-        ));
-
-        (text_tx, cancel, typed, batcher)
-    }
-
-    /// Issue: `whisrs cancel` during a streaming recording still typed the
-    /// phrase the backend flushed on shutdown. Once the cancel flag is set,
-    /// nothing further may reach the sink and the loop must exit.
-    #[tokio::test(start_paused = true)]
-    async fn typing_batcher_discards_text_after_cancel() {
-        let (text_tx, cancel, typed, batcher) = spawn_test_batcher();
-
-        // A delta before cancel is batched and typed.
-        text_tx.send("hello".to_string()).await.unwrap();
-        // Paused clock: sleeping past the 150ms batch window deterministically
-        // lets the batcher flush the batch to the sink.
-        tokio::time::sleep(std::time::Duration::from_millis(200)).await;
-        assert_eq!(*typed.lock().unwrap(), vec!["hello".to_string()]);
-
-        // Cancel, then simulate the backend flushing a trailing phrase.
-        cancel.store(true, Ordering::SeqCst);
-        text_tx
-            .send(" flushed trailing phrase".to_string())
-            .await
-            .unwrap();
-
-        // The sink never fires again and the loop exits, returning only the
-        // text typed before cancel.
-        let full_text = batcher.await.unwrap();
-        assert_eq!(full_text, "hello");
-        assert_eq!(*typed.lock().unwrap(), vec!["hello".to_string()]);
-
-        // Exiting dropped text_rx, so backend sends now fail fast.
-        assert!(text_tx.send("more".to_string()).await.is_err());
-    }
-
-    /// Cancel arriving while a batch is still being coalesced (inside the
-    /// 150ms window) must drop that batch before it reaches the sink.
-    #[tokio::test(start_paused = true)]
-    async fn typing_batcher_drops_batch_cancelled_during_window() {
-        let (text_tx, cancel, typed, batcher) = spawn_test_batcher();
-
-        text_tx.send("hello".to_string()).await.unwrap();
-        // Cancel mid-window, before the batch flushes at 150ms.
-        tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-        cancel.store(true, Ordering::SeqCst);
-
-        let full_text = batcher.await.unwrap();
-        assert_eq!(full_text, "");
-        assert!(typed.lock().unwrap().is_empty());
     }
 }
